@@ -11,10 +11,11 @@
 
 [package]
 edition = "2021"
-rust-version = "1.67"
+rust-version = "1.82"
 name = "icu_capi"
-version = "1.5.0"
+version = "2.0.0"
 authors = ["The ICU4X Project Developers"]
+build = false
 include = [
     "bindings/**/*",
     "!bindings/dart/**/*",
