@@ -5,9 +5,9 @@
 
 [package]
 edition = "2021"
-rust-version = "1.67"
+rust-version = "1.82"
 name = "icu_segmenter_data"
-version = "1.5.0"
+version = "2.0.0"
 authors = ["The ICU4X Project Developers"]
 include = [
     "data/**/*",
