@@ -1,16 +1,22 @@
-import { CodePointRangeIteratorResult } from "./CodePointRangeIteratorResult";
+// generated by diplomat-tool
+import type { CodePointRangeIteratorResult } from "./CodePointRangeIteratorResult"
+import type { pointer, codepoint } from "./diplomat-runtime.d.ts";
+
 
 /**
-
- * An iterator over code point ranges, produced by `ICU4XCodePointSetData` or one of the `ICU4XCodePointMapData` types
+ * An iterator over code point ranges, produced by `CodePointSetData` or
+ * one of the `CodePointMapData` types
  */
+
+
 export class CodePointRangeIterator {
+    get ffiValue(): pointer;
 
-  /**
 
-   * Advance the iterator by one and return the next range.
-
-   * If the iterator is out of items, `done` will be true
-   */
-  next(): CodePointRangeIteratorResult;
+    /**
+     * Advance the iterator by one and return the next range.
+     *
+     * If the iterator is out of items, `done` will be true
+     */
+    next(): CodePointRangeIteratorResult;
 }
