@@ -1,20 +1,15 @@
-/**
- * An error that occurred in Rust.
- */
-export class FFIError<E> extends Error {
-    error_value: E;
-}
-
-export type u8 = number;
-export type i8 = number;
-export type u16 = number;
-export type i16 = number;
-export type u32 = number;
-export type i32 = number;
-export type u64 = bigint;
-export type i64 = bigint;
-export type usize = number;
-export type isize = number;
-export type f32 = number;
-export type f64 = number;
-export type char = string;
+/// A [codepoint] is a Unicode code point, such as `a`, or `💡`.
+///
+/// The recommended way to obtain a `codepoint` is to create it from a
+/// `String`, which is conceptually a list of `codepoint`s. For example,
+/// `'a'.codePointAt(0)` is equal to the `char` `a`.
+///
+/// JS does not have a character/codepoint literal, so integer literals
+/// need to be used. For example the Unicode code point U+1F4A1, `💡`,
+/// can be represented by `0x1F4A1`. Note that only values in the ranges
+/// `0x0..0xD7FF` and `0xE000..0x10FFFF` (both inclusive) are Unicode
+/// code points, and hence valid `codepoint`s.
+///
+/// A `String` can be constructed from a `codepoint` using `String.fromCodePoint()`.
+export type codepoint = number;
+export type pointer = number;
