@@ -3,154 +3,207 @@
 // (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
 
 #[diplomat::bridge]
+#[diplomat::abi_rename = "icu4x_{0}_mv1"]
+#[diplomat::attr(auto, namespace = "icu4x")]
 pub mod ffi {
-    use crate::errors::ffi::ICU4XError;
-    use crate::locale::ffi::ICU4XLocale;
-    use crate::provider::ffi::ICU4XDataProvider;
     use alloc::boxed::Box;
-    #[allow(unused_imports)] // feature-specific
-    use icu_experimental::displaynames::{DisplayNamesOptions, Fallback, LanguageDisplay};
-    use icu_experimental::displaynames::{LocaleDisplayNamesFormatter, RegionDisplayNames};
-    use icu_locid::subtags::Region;
+
+    #[cfg(any(feature = "compiled_data", feature = "buffer_provider"))]
+    use crate::unstable::errors::ffi::DataError;
+    use crate::unstable::errors::ffi::LocaleParseError;
+    use crate::unstable::locale_core::ffi::Locale;
+    #[cfg(feature = "buffer_provider")]
+    use crate::unstable::provider::ffi::DataProvider;
+    use diplomat_runtime::DiplomatOption;
+
     use writeable::Writeable;
 
     #[diplomat::opaque]
-    #[diplomat::rust_link(icu::displaynames::LocaleDisplayNamesFormatter, Struct)]
-    pub struct ICU4XLocaleDisplayNamesFormatter(pub LocaleDisplayNamesFormatter);
+    #[diplomat::rust_link(icu::experimental::displaynames::LocaleDisplayNamesFormatter, Struct)]
+    pub struct LocaleDisplayNamesFormatter(
+        pub icu_experimental::displaynames::LocaleDisplayNamesFormatter,
+    );
 
     #[diplomat::opaque]
-    #[diplomat::rust_link(icu::displaynames::RegionDisplayNames, Struct)]
-    pub struct ICU4XRegionDisplayNames(pub RegionDisplayNames);
+    #[diplomat::rust_link(icu::experimental::displaynames::RegionDisplayNames, Struct)]
+    pub struct RegionDisplayNames(pub icu_experimental::displaynames::RegionDisplayNames);
 
-    #[diplomat::rust_link(icu::displaynames::options::DisplayNamesOptions, Struct)]
-    #[diplomat::attr(dart, rename = "DisplayNamesOptions")]
-    pub struct ICU4XDisplayNamesOptionsV1 {
+    #[diplomat::rust_link(icu::experimental::displaynames::options::DisplayNamesOptions, Struct)]
+    #[diplomat::attr(supports = non_exhaustive_structs, rename = "DisplayNamesOptions")]
+    pub struct DisplayNamesOptionsV1 {
         /// The optional formatting style to use for display name.
-        pub style: ICU4XDisplayNamesStyle,
+        pub style: DiplomatOption<DisplayNamesStyle>,
         /// The fallback return when the system does not have the
         /// requested display name, defaults to "code".
-        pub fallback: ICU4XDisplayNamesFallback,
+        pub fallback: DiplomatOption<DisplayNamesFallback>,
         /// The language display kind, defaults to "dialect".
-        pub language_display: ICU4XLanguageDisplay,
+        pub language_display: DiplomatOption<LanguageDisplay>,
     }
 
-    #[diplomat::rust_link(icu::displaynames::options::Style, Enum)]
-    pub enum ICU4XDisplayNamesStyle {
-        Auto,
+    #[diplomat::rust_link(icu::experimental::displaynames::options::Style, Enum)]
+    #[diplomat::enum_convert(icu_experimental::displaynames::Style, needs_wildcard)]
+    pub enum DisplayNamesStyle {
         Narrow,
         Short,
         Long,
         Menu,
     }
 
-    #[diplomat::rust_link(icu::displaynames::options::Fallback, Enum)]
-    #[diplomat::enum_convert(Fallback, needs_wildcard)]
-    pub enum ICU4XDisplayNamesFallback {
+    #[diplomat::rust_link(icu::experimental::displaynames::options::Fallback, Enum)]
+    #[diplomat::enum_convert(icu_experimental::displaynames::Fallback, needs_wildcard)]
+    pub enum DisplayNamesFallback {
         Code,
         None,
     }
 
-    #[diplomat::rust_link(icu::displaynames::options::LanguageDisplay, Enum)]
-    #[diplomat::enum_convert(LanguageDisplay, needs_wildcard)]
-    pub enum ICU4XLanguageDisplay {
+    #[diplomat::rust_link(icu::experimental::displaynames::options::LanguageDisplay, Enum)]
+    #[diplomat::enum_convert(icu_experimental::displaynames::LanguageDisplay, needs_wildcard)]
+    pub enum LanguageDisplay {
         Dialect,
         Standard,
     }
 
-    impl ICU4XLocaleDisplayNamesFormatter {
-        /// Creates a new `LocaleDisplayNamesFormatter` from locale data and an options bag.
-        #[diplomat::rust_link(icu::displaynames::LocaleDisplayNamesFormatter::try_new, FnInStruct)]
-        #[diplomat::attr(all(supports = constructors, supports = fallible_constructors), constructor)]
-        pub fn create(
-            provider: &ICU4XDataProvider,
-            locale: &ICU4XLocale,
-            options: ICU4XDisplayNamesOptionsV1,
-        ) -> Result<Box<ICU4XLocaleDisplayNamesFormatter>, ICU4XError> {
-            let locale = locale.to_datalocale();
-            let options = DisplayNamesOptions::from(options);
+    impl LocaleDisplayNamesFormatter {
+        /// Creates a new `LocaleDisplayNamesFormatter` from locale data and an options bag using compiled data.
+        #[diplomat::rust_link(
+            icu::experimental::displaynames::LocaleDisplayNamesFormatter::try_new,
+            FnInStruct
+        )]
+        #[diplomat::attr(all(supports = fallible_constructors, supports = non_exhaustive_structs), constructor)]
+        #[diplomat::attr(all(supports = fallible_constructors, supports = named_constructors, not(supports = non_exhaustive_structs)), named_constructor = "v1")]
+        #[diplomat::attr(supports = non_exhaustive_structs, rename = "create")]
+        #[cfg(feature = "compiled_data")]
+        pub fn create_v1(
+            locale: &Locale,
+            options: DisplayNamesOptionsV1,
+        ) -> Result<Box<LocaleDisplayNamesFormatter>, DataError> {
+            let prefs = (&locale.0).into();
+            let options = icu_experimental::displaynames::DisplayNamesOptions::from(options);
 
-            Ok(Box::new(ICU4XLocaleDisplayNamesFormatter(
-                call_constructor!(
-                    LocaleDisplayNamesFormatter::try_new,
-                    LocaleDisplayNamesFormatter::try_new_with_any_provider,
-                    LocaleDisplayNamesFormatter::try_new_with_buffer_provider,
-                    provider,
-                    &locale,
+            Ok(Box::new(LocaleDisplayNamesFormatter(
+                icu_experimental::displaynames::LocaleDisplayNamesFormatter::try_new(
+                    prefs, options,
+                )?,
+            )))
+        }
+
+        /// Creates a new `LocaleDisplayNamesFormatter` from locale data and an options bag using a particular data source.
+        #[diplomat::rust_link(
+            icu::experimental::displaynames::LocaleDisplayNamesFormatter::try_new,
+            FnInStruct
+        )]
+        #[diplomat::attr(supports = non_exhaustive_structs, rename = "create_with_provider")]
+        #[diplomat::attr(all(supports = fallible_constructors, supports = named_constructors, supports = non_exhaustive_structs), named_constructor = "with_provider")]
+        #[diplomat::attr(all(supports = fallible_constructors, supports = named_constructors, not(supports = non_exhaustive_structs)), named_constructor = "v1_with_provider")]
+        #[cfg(feature = "buffer_provider")]
+        pub fn create_v1_with_provider(
+            provider: &DataProvider,
+            locale: &Locale,
+            options: DisplayNamesOptionsV1,
+        ) -> Result<Box<LocaleDisplayNamesFormatter>, DataError> {
+            let prefs = (&locale.0).into();
+            let options = icu_experimental::displaynames::DisplayNamesOptions::from(options);
+
+            Ok(Box::new(LocaleDisplayNamesFormatter(
+                icu_experimental::displaynames::LocaleDisplayNamesFormatter::try_new_with_buffer_provider(provider.get()?, prefs,
                     options,
                 )?,
             )))
         }
 
         /// Returns the locale-specific display name of a locale.
-        #[diplomat::rust_link(icu::displaynames::LocaleDisplayNamesFormatter::of, FnInStruct)]
-        pub fn of(
-            &self,
-            locale: &ICU4XLocale,
-            write: &mut DiplomatWriteable,
-        ) -> Result<(), ICU4XError> {
-            self.0.of(&locale.0).write_to(write)?;
-            Ok(())
+        #[diplomat::rust_link(
+            icu::experimental::displaynames::LocaleDisplayNamesFormatter::of,
+            FnInStruct
+        )]
+        // Experimental, do not generate in demo:
+        #[diplomat::attr(demo_gen, disable)]
+        pub fn of(&self, locale: &Locale, write: &mut DiplomatWrite) {
+            let _infallible = self.0.of(&locale.0).write_to(write);
         }
     }
 
-    impl ICU4XRegionDisplayNames {
-        /// Creates a new `RegionDisplayNames` from locale data and an options bag.
-        #[diplomat::rust_link(icu::displaynames::RegionDisplayNames::try_new, FnInStruct)]
-        #[diplomat::attr(all(supports = constructors, supports = fallible_constructors), constructor)]
-        pub fn create(
-            provider: &ICU4XDataProvider,
-            locale: &ICU4XLocale,
-        ) -> Result<Box<ICU4XRegionDisplayNames>, ICU4XError> {
-            let locale = locale.to_datalocale();
-            Ok(Box::new(ICU4XRegionDisplayNames(call_constructor!(
-                RegionDisplayNames::try_new,
-                RegionDisplayNames::try_new_with_any_provider,
-                RegionDisplayNames::try_new_with_buffer_provider,
-                provider,
-                &locale,
-                Default::default()
-            )?)))
+    impl RegionDisplayNames {
+        /// Creates a new `RegionDisplayNames` from locale data and an options bag using compiled data.
+        #[diplomat::rust_link(
+            icu::experimental::displaynames::RegionDisplayNames::try_new,
+            FnInStruct
+        )]
+        #[diplomat::attr(all(supports = fallible_constructors, supports = non_exhaustive_structs), constructor)]
+        #[diplomat::attr(all(supports = fallible_constructors, supports = named_constructors, not(supports = non_exhaustive_structs)), named_constructor = "v1")]
+        #[diplomat::attr(supports = non_exhaustive_structs, rename = "create")]
+        #[cfg(feature = "compiled_data")]
+        pub fn create_v1(
+            locale: &Locale,
+            options: DisplayNamesOptionsV1,
+        ) -> Result<Box<RegionDisplayNames>, DataError> {
+            let prefs = (&locale.0).into();
+            let options = icu_experimental::displaynames::DisplayNamesOptions::from(options);
+            Ok(Box::new(RegionDisplayNames(
+                icu_experimental::displaynames::RegionDisplayNames::try_new(prefs, options)?,
+            )))
+        }
+
+        /// Creates a new `RegionDisplayNames` from locale data and an options bag using a particular data source.
+        #[diplomat::rust_link(
+            icu::experimental::displaynames::RegionDisplayNames::try_new,
+            FnInStruct
+        )]
+        #[diplomat::attr(supports = non_exhaustive_structs, rename = "create_with_provider")]
+        #[diplomat::attr(all(supports = fallible_constructors, supports = named_constructors, supports = non_exhaustive_structs), named_constructor = "with_provider")]
+        #[diplomat::attr(all(supports = fallible_constructors, supports = named_constructors, not(supports = non_exhaustive_structs)), named_constructor = "v1_with_provider")]
+        #[cfg(feature = "buffer_provider")]
+        pub fn create_v1_with_provider(
+            provider: &DataProvider,
+            locale: &Locale,
+            options: DisplayNamesOptionsV1,
+        ) -> Result<Box<RegionDisplayNames>, DataError> {
+            let prefs = (&locale.0).into();
+            let options = icu_experimental::displaynames::DisplayNamesOptions::from(options);
+            Ok(Box::new(RegionDisplayNames(
+                icu_experimental::displaynames::RegionDisplayNames::try_new_with_buffer_provider(
+                    provider.get()?,
+                    prefs,
+                    options,
+                )?,
+            )))
         }
 
         /// Returns the locale specific display name of a region.
         /// Note that the function returns an empty string in case the display name for a given
         /// region code is not found.
-        #[diplomat::rust_link(icu::displaynames::RegionDisplayNames::of, FnInStruct)]
+        #[diplomat::rust_link(icu::experimental::displaynames::RegionDisplayNames::of, FnInStruct)]
+        // Experimental, do not generate in demo:
+        #[diplomat::attr(demo_gen, disable)]
         pub fn of(
             &self,
             region: &DiplomatStr,
-            write: &mut DiplomatWriteable,
-        ) -> Result<(), ICU4XError> {
-            self.0
-                .of(Region::try_from_bytes(region)?)
+            write: &mut DiplomatWrite,
+        ) -> Result<(), LocaleParseError> {
+            let _infallible = self
+                .0
+                .of(icu_locale_core::subtags::Region::try_from_utf8(region)?)
                 .unwrap_or("")
-                .write_to(write)?;
+                .write_to(write);
             Ok(())
         }
     }
 }
 
-#[allow(unused_imports)] // feature-specific
-use icu_experimental::displaynames::{DisplayNamesOptions, Fallback, LanguageDisplay, Style};
-
-impl From<ffi::ICU4XDisplayNamesStyle> for Option<Style> {
-    fn from(style: ffi::ICU4XDisplayNamesStyle) -> Option<Style> {
-        match style {
-            ffi::ICU4XDisplayNamesStyle::Auto => None,
-            ffi::ICU4XDisplayNamesStyle::Narrow => Some(Style::Narrow),
-            ffi::ICU4XDisplayNamesStyle::Short => Some(Style::Short),
-            ffi::ICU4XDisplayNamesStyle::Long => Some(Style::Long),
-            ffi::ICU4XDisplayNamesStyle::Menu => Some(Style::Menu),
-        }
-    }
-}
-
-impl From<ffi::ICU4XDisplayNamesOptionsV1> for DisplayNamesOptions {
-    fn from(other: ffi::ICU4XDisplayNamesOptionsV1) -> DisplayNamesOptions {
-        let mut options = DisplayNamesOptions::default();
-        options.style = other.style.into();
-        options.fallback = other.fallback.into();
-        options.language_display = other.language_display.into();
+impl From<ffi::DisplayNamesOptionsV1> for icu_experimental::displaynames::DisplayNamesOptions {
+    fn from(
+        other: ffi::DisplayNamesOptionsV1,
+    ) -> icu_experimental::displaynames::DisplayNamesOptions {
+        let mut options = icu_experimental::displaynames::DisplayNamesOptions::default();
+        options.style = other.style.into_converted_option();
+        options.fallback = other
+            .fallback
+            .into_converted_option()
+            .unwrap_or(options.fallback);
+        options.language_display = other
+            .language_display
+            .into_converted_option()
+            .unwrap_or(options.language_display);
         options
     }
 }
