@@ -3,211 +3,172 @@
 // (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
 
 #[diplomat::bridge]
+#[diplomat::abi_rename = "icu4x_{0}_mv1"]
+#[diplomat::attr(auto, namespace = "icu4x")]
 pub mod ffi {
     use alloc::boxed::Box;
-    use icu_locid_transform::fallback::LocaleFallbackConfig;
-    use icu_locid_transform::fallback::LocaleFallbackIterator;
-    use icu_locid_transform::fallback::LocaleFallbackPriority;
-    use icu_locid_transform::fallback::LocaleFallbackerWithConfig;
-    use icu_locid_transform::LocaleFallbacker;
 
-    use crate::{
-        errors::ffi::ICU4XError, locale::ffi::ICU4XLocale, provider::ffi::ICU4XDataProvider,
-    };
+    use crate::unstable::locale_core::ffi::Locale;
+    #[cfg(feature = "buffer_provider")]
+    use crate::unstable::{errors::ffi::DataError, provider::ffi::DataProvider};
 
     /// An object that runs the ICU4X locale fallback algorithm.
     #[diplomat::opaque]
-    #[diplomat::rust_link(icu::locid_transform::fallback::LocaleFallbacker, Struct)]
-    #[diplomat::rust_link(
-        icu::locid_transform::fallback::LocaleFallbackerBorrowed,
-        Struct,
-        hidden
-    )]
-    pub struct ICU4XLocaleFallbacker(pub LocaleFallbacker);
+    #[diplomat::rust_link(icu::locale::fallback::LocaleFallbacker, Struct)]
+    #[diplomat::rust_link(icu::locale::fallback::LocaleFallbackerBorrowed, Struct, hidden)]
+    pub struct LocaleFallbacker(pub icu_locale::LocaleFallbacker);
 
     /// Priority mode for the ICU4X fallback algorithm.
-    #[diplomat::enum_convert(LocaleFallbackPriority, needs_wildcard)]
-    #[diplomat::rust_link(icu::locid_transform::fallback::LocaleFallbackPriority, Enum)]
+    #[diplomat::enum_convert(icu_locale::fallback::LocaleFallbackPriority, needs_wildcard)]
+    #[diplomat::rust_link(icu::locale::fallback::LocaleFallbackPriority, Enum)]
     #[diplomat::rust_link(
-        icu::locid_transform::fallback::LocaleFallbackPriority::const_default,
+        icu::locale::fallback::LocaleFallbackPriority::default,
         FnInEnum,
         hidden
     )]
-    pub enum ICU4XLocaleFallbackPriority {
+    pub enum LocaleFallbackPriority {
         Language = 0,
         Region = 1,
-        Collation = 2,
-    }
-
-    /// What additional data is required to load when performing fallback.
-    #[diplomat::rust_link(icu::locid_transform::fallback::LocaleFallbackSupplement, Enum)]
-    #[diplomat::rust_link(
-        icu::locid_transform::fallback::LocaleFallbackSupplement::const_default,
-        FnInEnum,
-        hidden
-    )]
-    pub enum ICU4XLocaleFallbackSupplement {
-        None = 0,
-        Collation = 1,
     }
 
     /// Collection of configurations for the ICU4X fallback algorithm.
-    #[diplomat::rust_link(icu::locid_transform::fallback::LocaleFallbackConfig, Struct)]
+    #[diplomat::rust_link(icu::locale::fallback::LocaleFallbackConfig, Struct)]
     #[diplomat::rust_link(
-        icu::locid_transform::fallback::LocaleFallbackConfig::const_default,
+        icu::locale::fallback::LocaleFallbackConfig::default,
         FnInStruct,
         hidden
     )]
-    pub struct ICU4XLocaleFallbackConfig<'a> {
+    pub struct LocaleFallbackConfig {
         /// Choice of priority mode.
-        pub priority: ICU4XLocaleFallbackPriority,
-        /// An empty string is considered `None`.
-        pub extension_key: &'a DiplomatStr,
-        /// Fallback supplement data key to customize fallback rules.
-        pub fallback_supplement: ICU4XLocaleFallbackSupplement,
+        pub priority: LocaleFallbackPriority,
     }
 
     /// An object that runs the ICU4X locale fallback algorithm with specific configurations.
     #[diplomat::opaque]
-    #[diplomat::rust_link(icu::locid_transform::fallback::LocaleFallbacker, Struct)]
-    #[diplomat::rust_link(icu::locid_transform::fallback::LocaleFallbackerWithConfig, Struct)]
-    pub struct ICU4XLocaleFallbackerWithConfig<'a>(pub LocaleFallbackerWithConfig<'a>);
+    #[diplomat::rust_link(icu::locale::fallback::LocaleFallbacker, Struct)]
+    #[diplomat::rust_link(icu::locale::fallback::LocaleFallbackerWithConfig, Struct)]
+    pub struct LocaleFallbackerWithConfig<'a>(
+        pub icu_locale::fallback::LocaleFallbackerWithConfig<'a>,
+    );
 
     /// An iterator over the locale under fallback.
     #[diplomat::opaque]
-    #[diplomat::rust_link(icu::locid_transform::fallback::LocaleFallbackIterator, Struct)]
-    pub struct ICU4XLocaleFallbackIterator<'a>(pub LocaleFallbackIterator<'a, 'a>);
+    #[diplomat::rust_link(icu::locale::fallback::LocaleFallbackIterator, Struct)]
+    pub struct LocaleFallbackIterator<'a>(pub icu_locale::fallback::LocaleFallbackIterator<'a>);
 
-    impl ICU4XLocaleFallbacker {
-        /// Creates a new `ICU4XLocaleFallbacker` from a data provider.
-        #[diplomat::rust_link(icu::locid_transform::fallback::LocaleFallbacker::new, FnInStruct)]
-        #[diplomat::attr(all(supports = constructors, supports = fallible_constructors), constructor)]
-        pub fn create(
-            provider: &ICU4XDataProvider,
-        ) -> Result<Box<ICU4XLocaleFallbacker>, ICU4XError> {
-            Ok(Box::new(ICU4XLocaleFallbacker(call_constructor!(
-                LocaleFallbacker::new [r => Ok(r.static_to_owned())],
-                LocaleFallbacker::try_new_with_any_provider,
-                LocaleFallbacker::try_new_with_buffer_provider,
-                provider,
-            )?)))
-        }
-
-        /// Creates a new `ICU4XLocaleFallbacker` without data for limited functionality.
+    impl LocaleFallbacker {
+        /// Creates a new `LocaleFallbacker` from compiled data.
+        #[diplomat::rust_link(icu::locale::fallback::LocaleFallbacker::new, FnInStruct)]
         #[diplomat::rust_link(
-            icu::locid_transform::fallback::LocaleFallbacker::new_without_data,
-            FnInStruct
-        )]
-        #[diplomat::attr(all(supports = constructors, supports = fallible_constructors, supports = named_constructors), named_constructor = "without_data")]
-        pub fn create_without_data() -> Box<ICU4XLocaleFallbacker> {
-            Box::new(ICU4XLocaleFallbacker(LocaleFallbacker::new_without_data()))
-        }
-
-        /// Associates this `ICU4XLocaleFallbacker` with configuration options.
-        #[diplomat::rust_link(
-            icu::locid_transform::fallback::LocaleFallbacker::for_config,
-            FnInStruct
-        )]
-        #[diplomat::rust_link(
-            icu::locid_transform::fallback::LocaleFallbackerBorrowed::for_config,
+            icu::locale::fallback::LocaleFallbackerBorrowed::new,
             FnInStruct,
             hidden
         )]
-        pub fn for_config<'a, 'temp>(
-            &'a self,
-            config: ICU4XLocaleFallbackConfig<'temp>,
-        ) -> Result<Box<ICU4XLocaleFallbackerWithConfig<'a>>, ICU4XError> {
-            Ok(Box::new(ICU4XLocaleFallbackerWithConfig(
-                self.0.for_config(LocaleFallbackConfig::try_from(config)?),
+        #[diplomat::attr(auto, constructor)]
+        #[cfg(feature = "compiled_data")]
+        pub fn create() -> Box<LocaleFallbacker> {
+            Box::new(LocaleFallbacker(
+                icu_locale::LocaleFallbacker::new().static_to_owned(),
+            ))
+        }
+
+        /// Creates a new `LocaleFallbacker` from a data provider.
+        #[diplomat::rust_link(icu::locale::fallback::LocaleFallbacker::new, FnInStruct)]
+        #[diplomat::rust_link(
+            icu::locale::fallback::LocaleFallbackerBorrowed::new,
+            FnInStruct,
+            hidden
+        )]
+        #[diplomat::attr(all(supports = fallible_constructors, supports = named_constructors), named_constructor = "with_provider")]
+        #[cfg(feature = "buffer_provider")]
+        pub fn create_with_provider(
+            provider: &DataProvider,
+        ) -> Result<Box<LocaleFallbacker>, DataError> {
+            Ok(Box::new(LocaleFallbacker(
+                icu_locale::LocaleFallbacker::try_new_with_buffer_provider(provider.get()?)?,
             )))
+        }
+
+        /// Creates a new `LocaleFallbacker` without data for limited functionality.
+        #[diplomat::rust_link(
+            icu::locale::fallback::LocaleFallbacker::new_without_data,
+            FnInStruct
+        )]
+        #[diplomat::attr(auto, named_constructor)]
+        pub fn without_data() -> Box<LocaleFallbacker> {
+            Box::new(LocaleFallbacker(
+                icu_locale::LocaleFallbacker::new_without_data(),
+            ))
+        }
+
+        /// Associates this `LocaleFallbacker` with configuration options.
+        #[diplomat::rust_link(icu::locale::fallback::LocaleFallbacker::for_config, FnInStruct)]
+        #[diplomat::rust_link(
+            icu::locale::fallback::LocaleFallbackerBorrowed::for_config,
+            FnInStruct,
+            hidden
+        )]
+        pub fn for_config<'a>(
+            &'a self,
+            config: LocaleFallbackConfig,
+        ) -> Box<LocaleFallbackerWithConfig<'a>> {
+            Box::new(LocaleFallbackerWithConfig(self.0.for_config({
+                let mut c = icu_locale::fallback::LocaleFallbackConfig::default();
+                c.priority = config.priority.into();
+                c
+            })))
         }
     }
 
-    impl<'a> ICU4XLocaleFallbackerWithConfig<'a> {
+    impl<'a> LocaleFallbackerWithConfig<'a> {
         /// Creates an iterator from a locale with each step of fallback.
+        #[diplomat::rust_link(icu::locale::fallback::LocaleFallbacker::fallback_for, FnInStruct)]
         #[diplomat::rust_link(
-            icu::locid_transform::fallback::LocaleFallbacker::fallback_for,
-            FnInStruct
-        )]
-        #[diplomat::rust_link(
-            icu::locid_transform::fallback::LocaleFallbackerBorrowed::fallback_for,
+            icu::locale::fallback::LocaleFallbackerBorrowed::fallback_for,
             FnInStruct,
             hidden
         )]
         #[diplomat::rust_link(
-            icu::locid_transform::fallback::LocaleFallbackerWithConfig::fallback_for,
+            icu::locale::fallback::LocaleFallbackerWithConfig::fallback_for,
             FnInStruct,
             hidden
         )]
         pub fn fallback_for_locale<'b: 'a, 'temp>(
             &'b self,
-            locale: &'temp ICU4XLocale,
-        ) -> Box<ICU4XLocaleFallbackIterator<'a>> {
-            Box::new(ICU4XLocaleFallbackIterator(
+            locale: &'temp Locale,
+        ) -> Box<LocaleFallbackIterator<'a>> {
+            Box::new(LocaleFallbackIterator(
                 self.0.fallback_for((&locale.0).into()),
             ))
         }
     }
 
-    impl<'a> ICU4XLocaleFallbackIterator<'a> {
-        /// Gets a snapshot of the current state of the locale.
+    impl<'a> LocaleFallbackIterator<'a> {
+        #[diplomat::attr(auto, iterator)]
         #[diplomat::rust_link(
-            icu::locid_transform::fallback::LocaleFallbackIterator::get,
-            FnInStruct
-        )]
-        #[diplomat::rust_link(
-            icu::locid_transform::fallback::LocaleFallbackIterator::take,
+            icu::locale::fallback::LocaleFallbackIterator::get,
             FnInStruct,
             hidden
         )]
-        #[diplomat::attr(*, disable)]
-        pub fn get(&self) -> Box<ICU4XLocale> {
-            Box::new(ICU4XLocale(self.0.get().clone().into_locale()))
-        }
-
-        /// Performs one step of the fallback algorithm, mutating the locale.
         #[diplomat::rust_link(
-            icu::locid_transform::fallback::LocaleFallbackIterator::step,
-            FnInStruct
+            icu::locale::fallback::LocaleFallbackIterator::step,
+            FnInStruct,
+            hidden
         )]
-        #[diplomat::attr(*, disable)]
-        pub fn step(&mut self) {
-            self.0.step();
-        }
-
-        /// A combination of `get` and `step`. Returns the value that `get` would return
-        /// and advances the iterator until hitting `und`.
-        #[diplomat::attr(supports = iterators, iterator)]
-        #[diplomat::skip_if_ast]
-        pub fn next(&mut self) -> Option<Box<ICU4XLocale>> {
-            let current = self.get();
-            if current.0 == icu_locid::Locale::UND {
+        #[diplomat::rust_link(
+            icu::locale::fallback::LocaleFallbackIterator::take,
+            FnInStruct,
+            hidden
+        )]
+        pub fn next(&mut self) -> Option<Box<Locale>> {
+            let current = self.0.get();
+            if current.is_unknown() {
                 None
             } else {
-                self.step();
-                Some(current)
+                let current = *current;
+                self.0.step();
+                Some(Box::new(Locale(current.into_locale())))
             }
         }
-    }
-}
-
-impl TryFrom<ffi::ICU4XLocaleFallbackConfig<'_>>
-    for icu_locid_transform::fallback::LocaleFallbackConfig
-{
-    type Error = crate::errors::ffi::ICU4XError;
-    fn try_from(other: ffi::ICU4XLocaleFallbackConfig) -> Result<Self, Self::Error> {
-        let mut result = Self::default();
-        result.priority = other.priority.into();
-        result.extension_key = match other.extension_key {
-            b"" => None,
-            s => Some(icu_locid::extensions::unicode::Key::try_from_bytes(s)?),
-        };
-        result.fallback_supplement = match other.fallback_supplement {
-            ffi::ICU4XLocaleFallbackSupplement::None => None,
-            ffi::ICU4XLocaleFallbackSupplement::Collation => {
-                Some(icu_locid_transform::fallback::LocaleFallbackSupplement::Collation)
-            }
-        };
-        Ok(result)
     }
 }
