@@ -3,102 +3,104 @@
 // (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
 
 #[diplomat::bridge]
+#[diplomat::abi_rename = "icu4x_{0}_mv1"]
+#[diplomat::attr(auto, namespace = "icu4x")]
 pub mod ffi {
-    use crate::{
-        errors::ffi::ICU4XError,
-        locale::ffi::ICU4XLocale,
-        locid_transform::ffi::ICU4XLocaleExpander,
-        provider::{ffi::ICU4XDataProvider, ICU4XDataProviderInner},
-    };
     use alloc::boxed::Box;
-    use icu_locid_transform::{Direction, LocaleDirectionality};
 
-    #[diplomat::rust_link(icu::locid_transform::Direction, Enum)]
-    pub enum ICU4XLocaleDirection {
+    #[cfg(feature = "buffer_provider")]
+    use crate::unstable::errors::ffi::DataError;
+    #[cfg(feature = "buffer_provider")]
+    use crate::unstable::provider::ffi::DataProvider;
+
+    use crate::unstable::locale_core::ffi::Locale;
+
+    #[diplomat::rust_link(icu::locale::Direction, Enum)]
+    pub enum LocaleDirection {
         LeftToRight,
         RightToLeft,
         Unknown,
     }
 
     #[diplomat::opaque]
-    #[diplomat::rust_link(icu::locid_transform::LocaleDirectionality, Struct)]
-    pub struct ICU4XLocaleDirectionality(pub LocaleDirectionality);
+    #[diplomat::rust_link(icu::locale::LocaleDirectionality, Struct)]
+    pub struct LocaleDirectionality(pub icu_locale::LocaleDirectionality);
 
-    impl ICU4XLocaleDirectionality {
-        /// Construct a new ICU4XLocaleDirectionality instance
-        #[diplomat::rust_link(icu::locid_transform::LocaleDirectionality::new, FnInStruct)]
-        #[diplomat::attr(all(supports = constructors, supports = fallible_constructors), constructor)]
-        pub fn create(
-            provider: &ICU4XDataProvider,
-        ) -> Result<Box<ICU4XLocaleDirectionality>, ICU4XError> {
-            Ok(Box::new(ICU4XLocaleDirectionality(call_constructor!(
-                LocaleDirectionality::new [r => Ok(r)],
-                LocaleDirectionality::try_new_with_any_provider,
-                LocaleDirectionality::try_new_with_buffer_provider,
-                provider,
-            )?)))
+    impl LocaleDirectionality {
+        /// Construct a new LocaleDirectionality instance using compiled data.
+        #[diplomat::rust_link(icu::locale::LocaleDirectionality::new_common, FnInStruct)]
+        #[diplomat::attr(supports = constructors, constructor)]
+        #[cfg(feature = "compiled_data")]
+        pub fn create_common() -> Box<LocaleDirectionality> {
+            Box::new(LocaleDirectionality(
+                icu_locale::LocaleDirectionality::new_common(),
+            ))
         }
 
-        /// Construct a new ICU4XLocaleDirectionality instance with a custom expander
+        /// Construct a new LocaleDirectionality instance using a particular data source.
+        #[diplomat::rust_link(icu::locale::LocaleDirectionality::new_common, FnInStruct)]
+        #[diplomat::attr(all(supports = fallible_constructors, supports = named_constructors), named_constructor = "with_provider")]
+        #[cfg(feature = "buffer_provider")]
+        pub fn create_common_with_provider(
+            provider: &DataProvider,
+        ) -> Result<Box<LocaleDirectionality>, DataError> {
+            Ok(Box::new(LocaleDirectionality(
+                icu_locale::LocaleDirectionality::try_new_common_with_buffer_provider(
+                    provider.get()?,
+                )?,
+            )))
+        }
+        /// Construct a new LocaleDirectionality instance using compiled data.
+        #[diplomat::rust_link(icu::locale::LocaleDirectionality::new_extended, FnInStruct)]
         #[diplomat::rust_link(
-            icu::locid_transform::LocaleDirectionality::new_with_expander,
-            FnInStruct
+            icu::locale::LocaleDirectionality::new_with_expander,
+            FnInStruct,
+            hidden
         )]
-        #[diplomat::attr(all(supports = constructors, supports = fallible_constructors, supports = named_constructors), named_constructor = "with_expander")]
-        pub fn create_with_expander(
-            provider: &ICU4XDataProvider,
-            expander: &ICU4XLocaleExpander,
-        ) -> Result<Box<ICU4XLocaleDirectionality>, ICU4XError> {
-            #[allow(unused_imports)]
-            use icu_provider::prelude::*;
-            Ok(Box::new(ICU4XLocaleDirectionality(match &provider.0 {
-                ICU4XDataProviderInner::Destroyed => Err(icu_provider::DataError::custom(
-                    "This provider has been destroyed",
-                ))?,
-                ICU4XDataProviderInner::Empty => {
-                    LocaleDirectionality::try_new_with_expander_unstable(
-                        &icu_provider_adapters::empty::EmptyDataProvider::new(),
-                        expander.0.clone(),
-                    )?
-                }
-                #[cfg(feature = "buffer_provider")]
-                ICU4XDataProviderInner::Buffer(buffer_provider) => {
-                    LocaleDirectionality::try_new_with_expander_unstable(
-                        &buffer_provider.as_deserializing(),
-                        expander.0.clone(),
-                    )?
-                }
-                #[cfg(feature = "compiled_data")]
-                ICU4XDataProviderInner::Compiled => {
-                    LocaleDirectionality::new_with_expander(expander.0.clone())
-                }
-            })))
+        #[diplomat::attr(auto, named_constructor = "extended")]
+        #[cfg(feature = "compiled_data")]
+        pub fn create_extended() -> Box<LocaleDirectionality> {
+            Box::new(LocaleDirectionality(
+                icu_locale::LocaleDirectionality::new_extended(),
+            ))
         }
 
-        #[diplomat::rust_link(icu::locid_transform::LocaleDirectionality::get, FnInStruct)]
-        #[diplomat::attr(supports = indexing, indexer)]
-        pub fn get(&self, locale: &ICU4XLocale) -> ICU4XLocaleDirection {
-            match self.0.get(&locale.0) {
-                Some(Direction::LeftToRight) => ICU4XLocaleDirection::LeftToRight,
-                Some(Direction::RightToLeft) => ICU4XLocaleDirection::RightToLeft,
-                _ => ICU4XLocaleDirection::Unknown,
+        /// Construct a new LocaleDirectionality instance using a particular data source.
+        #[diplomat::rust_link(icu::locale::LocaleDirectionality::new_extended, FnInStruct)]
+        #[diplomat::rust_link(
+            icu::locale::LocaleDirectionality::new_with_expander,
+            FnInStruct,
+            hidden
+        )]
+        #[diplomat::attr(all(supports = fallible_constructors, supports = named_constructors), named_constructor = "extended_with_provider")]
+        #[cfg(feature = "buffer_provider")]
+        pub fn create_extended_with_provider(
+            provider: &DataProvider,
+        ) -> Result<Box<LocaleDirectionality>, DataError> {
+            Ok(Box::new(LocaleDirectionality(
+                icu_locale::LocaleDirectionality::try_new_extended_with_buffer_provider(
+                    provider.get()?,
+                )?,
+            )))
+        }
+        #[diplomat::rust_link(icu::locale::LocaleDirectionality::get, FnInStruct)]
+        #[diplomat::attr(auto, indexer)]
+        pub fn get(&self, locale: &Locale) -> LocaleDirection {
+            match self.0.get(&locale.0.id) {
+                Some(icu_locale::Direction::LeftToRight) => LocaleDirection::LeftToRight,
+                Some(icu_locale::Direction::RightToLeft) => LocaleDirection::RightToLeft,
+                _ => LocaleDirection::Unknown,
             }
         }
 
-        #[diplomat::rust_link(
-            icu::locid_transform::LocaleDirectionality::is_left_to_right,
-            FnInStruct
-        )]
-        pub fn is_left_to_right(&self, locale: &ICU4XLocale) -> bool {
-            self.0.is_left_to_right(&locale.0)
+        #[diplomat::rust_link(icu::locale::LocaleDirectionality::is_left_to_right, FnInStruct)]
+        pub fn is_left_to_right(&self, locale: &Locale) -> bool {
+            self.0.is_left_to_right(&locale.0.id)
         }
 
-        #[diplomat::rust_link(
-            icu::locid_transform::LocaleDirectionality::is_right_to_left,
-            FnInStruct
-        )]
-        pub fn is_right_to_left(&self, locale: &ICU4XLocale) -> bool {
-            self.0.is_right_to_left(&locale.0)
+        #[diplomat::rust_link(icu::locale::LocaleDirectionality::is_right_to_left, FnInStruct)]
+        pub fn is_right_to_left(&self, locale: &Locale) -> bool {
+            self.0.is_right_to_left(&locale.0.id)
         }
     }
 }
