@@ -3,104 +3,168 @@
 // (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
 
 #[diplomat::bridge]
+#[diplomat::abi_rename = "icu4x_{0}_mv1"]
+#[diplomat::attr(auto, namespace = "icu4x")]
 pub mod ffi {
-    use crate::errors::ffi::ICU4XError;
-    use crate::provider::ffi::ICU4XDataProvider;
     use alloc::boxed::Box;
-    use core::convert::TryFrom;
-    use icu_segmenter::{
-        SentenceBreakIteratorLatin1, SentenceBreakIteratorPotentiallyIllFormedUtf8,
-        SentenceBreakIteratorUtf16, SentenceSegmenter,
-    };
+    use icu_segmenter::scaffold::{Latin1, PotentiallyIllFormedUtf8, Utf16};
+
+    #[cfg(feature = "buffer_provider")]
+    use crate::unstable::provider::ffi::DataProvider;
+    #[cfg(any(feature = "compiled_data", feature = "buffer_provider"))]
+    use crate::unstable::{errors::ffi::DataError, locale_core::ffi::Locale};
 
     #[diplomat::opaque]
     /// An ICU4X sentence-break segmenter, capable of finding sentence breakpoints in strings.
     #[diplomat::rust_link(icu::segmenter::SentenceSegmenter, Struct)]
-    pub struct ICU4XSentenceSegmenter(SentenceSegmenter);
+    #[diplomat::rust_link(icu::segmenter::SentenceSegmenterBorrowed, Struct, hidden)]
+    pub struct SentenceSegmenter(icu_segmenter::SentenceSegmenter);
 
     #[diplomat::opaque]
-    #[diplomat::rust_link(icu::segmenter::SentenceBreakIterator, Struct)]
-    #[diplomat::rust_link(
-        icu::segmenter::SentenceBreakIteratorPotentiallyIllFormedUtf8,
-        Typedef,
-        hidden
-    )]
-    #[diplomat::rust_link(icu::segmenter::SentenceBreakIteratorUtf8, Typedef, hidden)]
-    pub struct ICU4XSentenceBreakIteratorUtf8<'a>(
-        SentenceBreakIteratorPotentiallyIllFormedUtf8<'a, 'a>,
+    #[diplomat::rust_link(icu::segmenter::iterators::SentenceBreakIterator, Struct)]
+    pub struct SentenceBreakIteratorUtf8<'a>(
+        icu_segmenter::iterators::SentenceBreakIterator<'a, 'a, PotentiallyIllFormedUtf8>,
     );
 
     #[diplomat::opaque]
-    #[diplomat::rust_link(icu::segmenter::SentenceBreakIterator, Struct)]
-    #[diplomat::rust_link(icu::segmenter::SentenceBreakIteratorUtf16, Typedef, hidden)]
-    pub struct ICU4XSentenceBreakIteratorUtf16<'a>(SentenceBreakIteratorUtf16<'a, 'a>);
+    #[diplomat::rust_link(icu::segmenter::iterators::SentenceBreakIterator, Struct)]
+    pub struct SentenceBreakIteratorUtf16<'a>(
+        icu_segmenter::iterators::SentenceBreakIterator<'a, 'a, Utf16>,
+    );
 
     #[diplomat::opaque]
-    #[diplomat::rust_link(icu::segmenter::SentenceBreakIterator, Struct)]
-    #[diplomat::rust_link(icu::segmenter::SentenceBreakIteratorLatin1, Typedef, hidden)]
-    pub struct ICU4XSentenceBreakIteratorLatin1<'a>(SentenceBreakIteratorLatin1<'a, 'a>);
+    #[diplomat::rust_link(icu::segmenter::iterators::SentenceBreakIterator, Struct)]
+    pub struct SentenceBreakIteratorLatin1<'a>(
+        icu_segmenter::iterators::SentenceBreakIterator<'a, 'a, Latin1>,
+    );
 
-    impl ICU4XSentenceSegmenter {
-        /// Construct an [`ICU4XSentenceSegmenter`].
+    impl SentenceSegmenter {
+        /// Construct a [`SentenceSegmenter`] using compiled data. This does not assume any content locale.
         #[diplomat::rust_link(icu::segmenter::SentenceSegmenter::new, FnInStruct)]
-        #[diplomat::attr(all(supports = constructors, supports = fallible_constructors), constructor)]
-        pub fn create(
-            provider: &ICU4XDataProvider,
-        ) -> Result<Box<ICU4XSentenceSegmenter>, ICU4XError> {
-            Ok(Box::new(ICU4XSentenceSegmenter(call_constructor!(
-                SentenceSegmenter::new [r => Ok(r)],
-                SentenceSegmenter::try_new_with_any_provider,
-                SentenceSegmenter::try_new_with_buffer_provider,
-                provider,
-            )?)))
+        #[diplomat::rust_link(
+            icu::segmenter::options::SentenceBreakInvariantOptions,
+            Struct,
+            hidden
+        )]
+        #[diplomat::attr(auto, constructor)]
+        #[cfg(feature = "compiled_data")]
+        pub fn create() -> Box<SentenceSegmenter> {
+            Box::new(SentenceSegmenter(
+                icu_segmenter::SentenceSegmenter::new(Default::default()).static_to_owned(),
+            ))
+        }
+        /// Construct a [`SentenceSegmenter`] for content known to be of a given locale, using compiled data.
+        #[diplomat::rust_link(icu::segmenter::SentenceSegmenter::try_new, FnInStruct, hidden)]
+        #[diplomat::rust_link(icu::segmenter::options::SentenceBreakOptions, Struct, hidden)]
+        #[diplomat::attr(all(supports = fallible_constructors, supports = named_constructors), named_constructor = "with_content_locale")]
+        #[cfg(feature = "compiled_data")]
+        pub fn create_with_content_locale(
+            locale: &Locale,
+        ) -> Result<Box<SentenceSegmenter>, DataError> {
+            Ok(Box::new(SentenceSegmenter(
+                icu_segmenter::SentenceSegmenter::try_new(locale.into())?,
+            )))
+        }
+
+        /// Construct a [`SentenceSegmenter`]  for content known to be of a given locale, using a particular data source.
+        #[diplomat::rust_link(icu::segmenter::SentenceSegmenter::try_new, FnInStruct, hidden)]
+        #[diplomat::attr(all(supports = fallible_constructors, supports = named_constructors), named_constructor = "with_content_locale_and_provider")]
+        #[cfg(feature = "buffer_provider")]
+        pub fn create_with_content_locale_and_provider(
+            provider: &DataProvider,
+            locale: &Locale,
+        ) -> Result<Box<SentenceSegmenter>, DataError> {
+            Ok(Box::new(SentenceSegmenter(
+                icu_segmenter::SentenceSegmenter::try_new_with_buffer_provider(
+                    provider.get()?,
+                    locale.into(),
+                )?,
+            )))
         }
 
         /// Segments a string.
         ///
         /// Ill-formed input is treated as if errors had been replaced with REPLACEMENT CHARACTERs according
         /// to the WHATWG Encoding Standard.
-        #[diplomat::rust_link(icu::segmenter::SentenceSegmenter::segment_utf8, FnInStruct)]
-        #[diplomat::rust_link(icu::segmenter::SentenceSegmenter::segment_str, FnInStruct, hidden)]
-        #[diplomat::attr(dart, disable)]
+        #[diplomat::rust_link(icu::segmenter::SentenceSegmenterBorrowed::segment_utf8, FnInStruct)]
+        #[diplomat::rust_link(
+            icu::segmenter::SentenceSegmenterBorrowed::segment_str,
+            FnInStruct,
+            hidden
+        )]
+        #[diplomat::attr(not(supports = utf8_strings), disable)]
+        #[diplomat::attr(*, rename = "segment")]
         pub fn segment_utf8<'a>(
             &'a self,
             input: &'a DiplomatStr,
-        ) -> Box<ICU4XSentenceBreakIteratorUtf8<'a>> {
-            Box::new(ICU4XSentenceBreakIteratorUtf8(self.0.segment_utf8(input)))
+        ) -> Box<SentenceBreakIteratorUtf8<'a>> {
+            Box::new(SentenceBreakIteratorUtf8(
+                self.0.as_borrowed().segment_utf8(input),
+            ))
         }
 
         /// Segments a string.
         ///
         /// Ill-formed input is treated as if errors had been replaced with REPLACEMENT CHARACTERs according
         /// to the WHATWG Encoding Standard.
-        #[diplomat::rust_link(icu::segmenter::SentenceSegmenter::segment_utf16, FnInStruct)]
-        #[diplomat::attr(dart, rename = "segment")]
+        #[diplomat::rust_link(icu::segmenter::SentenceSegmenterBorrowed::segment_utf16, FnInStruct)]
+        #[diplomat::attr(not(supports = utf8_strings), rename = "segment")]
+        #[diplomat::attr(supports = utf8_strings, rename = "segment16")]
         pub fn segment_utf16<'a>(
             &'a self,
             input: &'a DiplomatStr16,
-        ) -> Box<ICU4XSentenceBreakIteratorUtf16<'a>> {
-            Box::new(ICU4XSentenceBreakIteratorUtf16(self.0.segment_utf16(input)))
+        ) -> Box<SentenceBreakIteratorUtf16<'a>> {
+            Box::new(SentenceBreakIteratorUtf16(
+                self.0.as_borrowed().segment_utf16(input),
+            ))
         }
 
         /// Segments a Latin-1 string.
-        #[diplomat::rust_link(icu::segmenter::SentenceSegmenter::segment_latin1, FnInStruct)]
-        #[diplomat::attr(dart, disable)]
+        #[diplomat::rust_link(
+            icu::segmenter::SentenceSegmenterBorrowed::segment_latin1,
+            FnInStruct
+        )]
+        #[diplomat::attr(not(supports = utf8_strings), disable)]
         pub fn segment_latin1<'a>(
             &'a self,
             input: &'a [u8],
-        ) -> Box<ICU4XSentenceBreakIteratorLatin1<'a>> {
-            Box::new(ICU4XSentenceBreakIteratorLatin1(
-                self.0.segment_latin1(input),
+        ) -> Box<SentenceBreakIteratorLatin1<'a>> {
+            Box::new(SentenceBreakIteratorLatin1(
+                self.0.as_borrowed().segment_latin1(input),
             ))
         }
     }
 
-    impl<'a> ICU4XSentenceBreakIteratorUtf8<'a> {
+    impl<'a> SentenceBreakIteratorUtf8<'a> {
         /// Finds the next breakpoint. Returns -1 if at the end of the string or if the index is
         /// out of range of a 32-bit signed integer.
-        #[diplomat::rust_link(icu::segmenter::SentenceBreakIterator::next, FnInStruct)]
+        #[diplomat::rust_link(icu::segmenter::iterators::SentenceBreakIterator::next, FnInStruct)]
+        pub fn next(&mut self) -> i32 {
+            self.0
+                .next()
+                .and_then(|u| i32::try_from(u).ok())
+                .unwrap_or(-1)
+        }
+    }
+
+    impl<'a> SentenceBreakIteratorUtf16<'a> {
+        /// Finds the next breakpoint. Returns -1 if at the end of the string or if the index is
+        /// out of range of a 32-bit signed integer.
+        #[diplomat::rust_link(icu::segmenter::iterators::SentenceBreakIterator::next, FnInStruct)]
+        pub fn next(&mut self) -> i32 {
+            self.0
+                .next()
+                .and_then(|u| i32::try_from(u).ok())
+                .unwrap_or(-1)
+        }
+    }
+
+    impl<'a> SentenceBreakIteratorLatin1<'a> {
+        /// Finds the next breakpoint. Returns -1 if at the end of the string or if the index is
+        /// out of range of a 32-bit signed integer.
+        #[diplomat::rust_link(icu::segmenter::iterators::SentenceBreakIterator::next, FnInStruct)]
         #[diplomat::rust_link(
-            icu::segmenter::SentenceBreakIterator::Item,
+            icu::segmenter::iterators::SentenceBreakIterator::Item,
             AssociatedTypeInStruct,
             hidden
         )]
@@ -111,38 +175,14 @@ pub mod ffi {
                 .unwrap_or(-1)
         }
     }
+}
 
-    impl<'a> ICU4XSentenceBreakIteratorUtf16<'a> {
-        /// Finds the next breakpoint. Returns -1 if at the end of the string or if the index is
-        /// out of range of a 32-bit signed integer.
-        #[diplomat::rust_link(icu::segmenter::SentenceBreakIterator::next, FnInStruct)]
-        #[diplomat::rust_link(
-            icu::segmenter::SentenceBreakIterator::Item,
-            AssociatedTypeInStruct,
-            hidden
-        )]
-        pub fn next(&mut self) -> i32 {
-            self.0
-                .next()
-                .and_then(|u| i32::try_from(u).ok())
-                .unwrap_or(-1)
-        }
-    }
-
-    impl<'a> ICU4XSentenceBreakIteratorLatin1<'a> {
-        /// Finds the next breakpoint. Returns -1 if at the end of the string or if the index is
-        /// out of range of a 32-bit signed integer.
-        #[diplomat::rust_link(icu::segmenter::SentenceBreakIterator::next, FnInStruct)]
-        #[diplomat::rust_link(
-            icu::segmenter::SentenceBreakIterator::Item,
-            AssociatedTypeInStruct,
-            hidden
-        )]
-        pub fn next(&mut self) -> i32 {
-            self.0
-                .next()
-                .and_then(|u| i32::try_from(u).ok())
-                .unwrap_or(-1)
-        }
+impl<'a> From<&'a crate::unstable::locale_core::ffi::Locale>
+    for icu_segmenter::options::SentenceBreakOptions<'a>
+{
+    fn from(other: &'a crate::unstable::locale_core::ffi::Locale) -> Self {
+        let mut options = icu_segmenter::options::SentenceBreakOptions::default();
+        options.content_locale = Some(&other.0.id);
+        options
     }
 }
