@@ -2,378 +2,249 @@
 // called LICENSE at the top level of the ICU4X source tree
 // (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
 
-use icu_timezone::CustomTimeZone;
+use ffi::TimeZoneInfo;
 
 #[diplomat::bridge]
+#[diplomat::abi_rename = "icu4x_{0}_mv1"]
+#[diplomat::attr(auto, namespace = "icu4x")]
 pub mod ffi {
-    use crate::errors::ffi::ICU4XError;
     use alloc::boxed::Box;
-    use core::fmt::Write;
-    use icu_timezone::CustomTimeZone;
-    use icu_timezone::GmtOffset;
-    use icu_timezone::ZoneVariant;
+
+    use crate::unstable::{
+        date::ffi::IsoDate,
+        datetime::ffi::IsoDateTime,
+        time::ffi::Time,
+        variant_offset::ffi::{UtcOffset, VariantOffsetsCalculator},
+    };
 
     #[diplomat::opaque]
-    #[diplomat::rust_link(icu::timezone::CustomTimeZone, Struct)]
-    pub struct ICU4XCustomTimeZone(pub CustomTimeZone);
+    #[diplomat::rust_link(icu::time::TimeZone, Struct)]
+    pub struct TimeZone(pub(crate) icu_time::TimeZone);
 
-    impl ICU4XCustomTimeZone {
-        /// Creates a time zone from an offset string.
-        #[diplomat::rust_link(icu::timezone::CustomTimeZone::from_str, FnInStruct)]
-        #[diplomat::rust_link(icu::timezone::CustomTimeZone::try_from_bytes, FnInStruct, hidden)]
-        #[diplomat::rust_link(icu::timezone::GmtOffset::from_str, FnInStruct, hidden)]
-        #[diplomat::rust_link(icu::timezone::GmtOffset::try_from_bytes, FnInStruct, hidden)]
-        #[diplomat::attr(all(supports = constructors, supports = fallible_constructors, supports = named_constructors), named_constructor = "from_string")]
-        pub fn create_from_string(s: &DiplomatStr) -> Result<Box<ICU4XCustomTimeZone>, ICU4XError> {
-            Ok(Box::new(ICU4XCustomTimeZone::from(
-                CustomTimeZone::try_from_bytes(s)?,
-            )))
+    impl TimeZone {
+        /// The unknown time zone.
+        #[diplomat::rust_link(icu::time::TimeZoneInfo::unknown, FnInStruct)]
+        #[diplomat::rust_link(icu::time::TimeZone::unknown, FnInStruct, hidden)]
+        #[diplomat::attr(auto, named_constructor)]
+        pub fn unknown() -> Box<TimeZone> {
+            Box::new(TimeZone(icu_time::TimeZone::UNKNOWN))
         }
 
-        /// Creates a time zone with no information.
-        #[diplomat::rust_link(icu::timezone::CustomTimeZone::new_empty, FnInStruct)]
-        #[diplomat::attr(all(supports = constructors, supports = fallible_constructors, supports = named_constructors), named_constructor = "empty")]
-        pub fn create_empty() -> Box<ICU4XCustomTimeZone> {
-            Box::new(CustomTimeZone::new_empty().into())
+        /// Whether the time zone is the unknown zone.
+        #[diplomat::rust_link(icu::time::TimeZone::is_unknown, FnInStruct)]
+        pub fn is_unknown(&self) -> bool {
+            self.0.is_unknown()
         }
 
-        /// Creates a time zone for UTC.
-        #[diplomat::rust_link(icu::timezone::CustomTimeZone::utc, FnInStruct)]
-        #[diplomat::rust_link(icu::timezone::GmtOffset::utc, FnInStruct, hidden)]
-        #[diplomat::attr(all(supports = constructors, supports = fallible_constructors, supports = named_constructors), named_constructor = "utc")]
-        pub fn create_utc() -> Box<ICU4XCustomTimeZone> {
-            Box::new(CustomTimeZone::utc().into())
-        }
-
-        /// Sets the `gmt_offset` field from offset seconds.
+        /// Creates a time zone from a BCP-47 string.
         ///
-        /// Errors if the offset seconds are out of range.
-        #[diplomat::rust_link(icu::timezone::GmtOffset::try_from_offset_seconds, FnInStruct)]
-        #[diplomat::rust_link(icu::timezone::GmtOffset, Struct, compact)]
+        /// Returns the unknown time zone if the string is not a valid BCP-47 subtag.
+        #[diplomat::rust_link(icu::time::TimeZone, Struct, compact)]
+        #[diplomat::attr(auto, named_constructor = "from_bcp47")]
+        #[diplomat::demo(default_constructor)]
+        pub fn create_from_bcp47(id: &DiplomatStr) -> Box<Self> {
+            icu_locale_core::subtags::Subtag::try_from_utf8(id)
+                .map(icu_time::TimeZone)
+                .map(TimeZone)
+                .map(Box::new)
+                .unwrap_or(Self::unknown())
+        }
+
+        #[diplomat::rust_link(icu::time::TimeZone::with_offset, FnInStruct)]
+        pub fn with_offset(&self, offset: &UtcOffset) -> Box<TimeZoneInfo> {
+            Box::new(self.0.with_offset(Some(offset.0)).into())
+        }
+
+        #[diplomat::rust_link(icu::time::TimeZone::without_offset, FnInStruct)]
+        pub fn without_offset(&self) -> Box<TimeZoneInfo> {
+            Box::new(self.0.without_offset().into())
+        }
+    }
+
+    #[diplomat::enum_convert(icu_time::zone::TimeZoneVariant, needs_wildcard)]
+    pub enum TimeZoneVariant {
+        Standard,
+        Daylight,
+    }
+
+    impl TimeZoneVariant {
+        /// Sets the `variant` field to "daylight" time.
+        #[diplomat::rust_link(icu::time::zone::TimeZoneVariant::from_rearguard_isdst, FnInEnum)]
+        #[diplomat::rust_link(icu::time::TimeZoneInfo::with_variant, FnInStruct)]
+        #[diplomat::rust_link(icu::time::zone::TimeZoneVariant, Enum, compact)]
+        pub fn from_rearguard_isdst(&mut self, isdst: bool) -> Self {
+            icu_time::zone::TimeZoneVariant::from_rearguard_isdst(isdst).into()
+        }
+    }
+
+    #[diplomat::opaque]
+    #[diplomat::rust_link(icu::time::TimeZoneInfo, Struct)]
+    #[diplomat::rust_link(icu::time::zone::models::AtTime, Struct, hidden)]
+    #[diplomat::rust_link(icu::time::zone::models::Base, Struct, hidden)]
+    #[diplomat::rust_link(icu::time::zone::models::Full, Struct, hidden)]
+    pub struct TimeZoneInfo {
+        pub(crate) id: icu_time::TimeZone,
+        pub(crate) offset: Option<icu_time::zone::UtcOffset>,
+        pub(crate) variant: Option<icu_time::zone::TimeZoneVariant>,
+        pub(crate) zone_name_timestamp: Option<icu_time::zone::ZoneNameTimestamp>,
+    }
+
+    impl TimeZoneInfo {
+        /// Creates a time zone for UTC (Coordinated Universal Time).
+        #[diplomat::rust_link(icu::time::TimeZoneInfo::utc, FnInStruct)]
+        #[diplomat::rust_link(icu::time::zone::UtcOffset::zero, FnInStruct, hidden)]
+        #[diplomat::attr(auto, named_constructor)]
+        pub fn utc() -> Box<TimeZoneInfo> {
+            Box::new(icu_time::TimeZoneInfo::utc().into())
+        }
+
+        /// Creates a time zone info from parts.
+        #[diplomat::attr(auto, constructor)]
+        pub fn from_parts(
+            id: &TimeZone,
+            offset: Option<&UtcOffset>,
+            variant: Option<TimeZoneVariant>,
+        ) -> Box<TimeZoneInfo> {
+            Box::new(Self {
+                id: id.0,
+                offset: offset.map(|o| o.0),
+                variant: variant.map(Into::into),
+                zone_name_timestamp: None,
+            })
+        }
+
+        #[diplomat::rust_link(icu::time::TimeZoneInfo::id, FnInStruct)]
+        pub fn id(&self) -> Box<TimeZone> {
+            Box::new(TimeZone(self.id))
+        }
+
+        /// Sets the datetime at which to interpret the time zone
+        /// for display name lookup.
+        ///
+        /// Notes:
+        ///
+        /// - If not set, the formatting datetime is used if possible.
+        /// - The constraints are the same as with `ZoneNameTimestamp` in Rust.
+        /// - Set to year 1000 or 9999 for a reference far in the past or future.
+        #[diplomat::rust_link(icu::time::TimeZoneInfo::at_date_time_iso, FnInStruct)]
+        #[diplomat::rust_link(icu::time::zone::ZoneNameTimestamp, Struct, compact)]
         #[diplomat::rust_link(
-            icu::timezone::GmtOffset::from_offset_seconds_unchecked,
+            icu::time::TimeZoneInfo::with_zone_name_timestamp,
             FnInStruct,
             hidden
         )]
-        #[diplomat::rust_link(icu::timezone::CustomTimeZone::new_with_offset, FnInStruct, hidden)]
-        pub fn try_set_gmt_offset_seconds(
-            &mut self,
-            offset_seconds: i32,
-        ) -> Result<(), ICU4XError> {
-            self.0.gmt_offset = Some(GmtOffset::try_from_offset_seconds(offset_seconds)?);
-            Ok(())
-        }
-
-        /// Clears the `gmt_offset` field.
-        #[diplomat::rust_link(icu::timezone::GmtOffset::offset_seconds, FnInStruct)]
-        #[diplomat::rust_link(icu::timezone::GmtOffset, Struct, compact)]
-        pub fn clear_gmt_offset(&mut self) {
-            self.0.gmt_offset.take();
-        }
-
-        /// Returns the value of the `gmt_offset` field as offset seconds.
-        ///
-        /// Errors if the `gmt_offset` field is empty.
-        #[diplomat::rust_link(icu::timezone::GmtOffset::offset_seconds, FnInStruct)]
-        #[diplomat::rust_link(icu::timezone::GmtOffset, Struct, compact)]
-        #[diplomat::attr(supports = accessors, getter)]
-        pub fn gmt_offset_seconds(&self) -> Result<i32, ICU4XError> {
-            self.0
-                .gmt_offset
-                .ok_or(ICU4XError::TimeZoneMissingInputError)
-                .map(GmtOffset::offset_seconds)
-        }
-
-        /// Returns whether the `gmt_offset` field is positive.
-        ///
-        /// Errors if the `gmt_offset` field is empty.
-        #[diplomat::rust_link(icu::timezone::GmtOffset::is_positive, FnInStruct)]
-        #[diplomat::attr(supports = accessors, getter)]
-        pub fn is_gmt_offset_positive(&self) -> Result<bool, ICU4XError> {
-            self.0
-                .gmt_offset
-                .ok_or(ICU4XError::TimeZoneMissingInputError)
-                .map(GmtOffset::is_positive)
-        }
-
-        /// Returns whether the `gmt_offset` field is zero.
-        ///
-        /// Errors if the `gmt_offset` field is empty (which is not the same as zero).
-        #[diplomat::rust_link(icu::timezone::GmtOffset::is_zero, FnInStruct)]
-        #[diplomat::attr(supports = accessors, getter)]
-        pub fn is_gmt_offset_zero(&self) -> Result<bool, ICU4XError> {
-            self.0
-                .gmt_offset
-                .ok_or(ICU4XError::TimeZoneMissingInputError)
-                .map(GmtOffset::is_zero)
-        }
-
-        /// Returns whether the `gmt_offset` field has nonzero minutes.
-        ///
-        /// Errors if the `gmt_offset` field is empty.
-        #[diplomat::rust_link(icu::timezone::GmtOffset::has_minutes, FnInStruct)]
-        #[diplomat::attr(supports = accessors, getter)]
-        pub fn gmt_offset_has_minutes(&self) -> Result<bool, ICU4XError> {
-            self.0
-                .gmt_offset
-                .ok_or(ICU4XError::TimeZoneMissingInputError)
-                .map(GmtOffset::has_minutes)
-        }
-
-        /// Returns whether the `gmt_offset` field has nonzero seconds.
-        ///
-        /// Errors if the `gmt_offset` field is empty.
-        #[diplomat::rust_link(icu::timezone::GmtOffset::has_seconds, FnInStruct)]
-        #[diplomat::attr(supports = accessors, getter)]
-        pub fn gmt_offset_has_seconds(&self) -> Result<bool, ICU4XError> {
-            self.0
-                .gmt_offset
-                .ok_or(ICU4XError::TimeZoneMissingInputError)
-                .map(GmtOffset::has_seconds)
-        }
-
-        /// Sets the `time_zone_id` field from a BCP-47 string.
-        ///
-        /// Errors if the string is not a valid BCP-47 time zone ID.
-        #[diplomat::rust_link(icu::timezone::CustomTimeZone::time_zone_id, StructField)]
-        #[diplomat::rust_link(icu::timezone::TimeZoneBcp47Id, Struct, compact)]
-        #[diplomat::rust_link(icu::timezone::TimeZoneBcp47Id::from_str, FnInStruct, hidden)]
-        #[diplomat::rust_link(icu::timezone::TimeZoneBcp47Id::deref, FnInStruct, hidden)]
         #[diplomat::rust_link(
-            icu::timezone::TimeZoneBcp47Id::Target,
-            AssociatedTypeInStruct,
+            icu::time::zone::ZoneNameTimestamp::from_date_time_iso,
+            FnInStruct,
             hidden
         )]
-        pub fn try_set_time_zone_id(&mut self, id: &DiplomatStr) -> Result<(), ICU4XError> {
-            self.0.time_zone_id = Some(icu_timezone::TimeZoneBcp47Id(
-                tinystr::TinyAsciiStr::from_bytes(id)
-                    .map_err(|_| ICU4XError::TimeZoneInvalidIdError)?,
-            ));
-            Ok(())
-        }
-
-        /// Sets the `time_zone_id` field from an IANA string by looking up
-        /// the corresponding BCP-47 string.
-        ///
-        /// Errors if the string is not a valid BCP-47 time zone ID.
-        #[diplomat::rust_link(icu::timezone::IanaToBcp47MapperBorrowed::get, FnInStruct)]
-        pub fn try_set_iana_time_zone_id(
-            &mut self,
-            mapper: &crate::iana_bcp47_mapper::ffi::ICU4XIanaToBcp47Mapper,
-            id: &DiplomatStr,
-        ) -> Result<(), ICU4XError> {
-            let id = core::str::from_utf8(id).map_err(|_| ICU4XError::TimeZoneInvalidIdError)?;
-            self.0.time_zone_id = Some(
-                mapper
-                    .0
-                    .as_borrowed()
-                    .get(id)
-                    .ok_or(ICU4XError::TimeZoneInvalidIdError)?,
-            );
-            Ok(())
-        }
-
-        // *** TODO: in 2.0 please replace try_set_iana_time_zone_id with try_set_iana_time_zone_id_2 ***
-
-        /// Sets the `time_zone_id` field from an IANA string by looking up
-        /// the corresponding BCP-47 string.
-        ///
-        /// Errors if the string is not a valid BCP-47 time zone ID.
-        pub fn try_set_iana_time_zone_id_2(
-            &mut self,
-            mapper: &crate::timezone_mapper::ffi::ICU4XTimeZoneIdMapper,
-            id: &DiplomatStr,
-        ) -> Result<(), ICU4XError> {
-            self.0.time_zone_id = Some(
-                mapper
-                    .0
-                    .as_borrowed()
-                    .iana_bytes_to_bcp47(id)
-                    .ok_or(ICU4XError::TimeZoneInvalidIdError)?,
-            );
-            Ok(())
-        }
-
-        /// Clears the `time_zone_id` field.
-        #[diplomat::rust_link(icu::timezone::CustomTimeZone::time_zone_id, StructField)]
-        #[diplomat::rust_link(icu::timezone::TimeZoneBcp47Id, Struct, compact)]
-        pub fn clear_time_zone_id(&mut self) {
-            self.0.time_zone_id.take();
-        }
-
-        /// Writes the value of the `time_zone_id` field as a string.
-        ///
-        /// Errors if the `time_zone_id` field is empty.
-        #[diplomat::rust_link(icu::timezone::CustomTimeZone::time_zone_id, StructField)]
-        #[diplomat::rust_link(icu::timezone::TimeZoneBcp47Id, Struct, compact)]
-        #[diplomat::attr(supports = accessors, getter)]
-        pub fn time_zone_id(
-            &self,
-            write: &mut diplomat_runtime::DiplomatWriteable,
-        ) -> Result<(), ICU4XError> {
-            write.write_str(
-                self.0
-                    .time_zone_id
-                    .ok_or(ICU4XError::TimeZoneMissingInputError)?
-                    .0
-                    .as_str(),
-            )?;
-            Ok(())
-        }
-
-        /// Sets the `metazone_id` field from a string.
-        ///
-        /// Errors if the string is not a valid BCP-47 metazone ID.
-        #[diplomat::rust_link(icu::timezone::CustomTimeZone::metazone_id, StructField)]
-        #[diplomat::rust_link(icu::timezone::MetazoneId, Struct, compact)]
-        #[diplomat::rust_link(icu::timezone::MetazoneId::from_str, FnInStruct, hidden)]
-        pub fn try_set_metazone_id(&mut self, id: &DiplomatStr) -> Result<(), ICU4XError> {
-            self.0.metazone_id = Some(icu_timezone::MetazoneId(
-                tinystr::TinyAsciiStr::from_bytes(id)
-                    .map_err(|_| ICU4XError::TimeZoneInvalidIdError)?,
-            ));
-            Ok(())
-        }
-
-        /// Clears the `metazone_id` field.
-        #[diplomat::rust_link(icu::timezone::CustomTimeZone::metazone_id, StructField)]
-        #[diplomat::rust_link(icu::timezone::MetazoneId, Struct, compact)]
-        pub fn clear_metazone_id(&mut self) {
-            self.0.metazone_id.take();
-        }
-
-        /// Writes the value of the `metazone_id` field as a string.
-        ///
-        /// Errors if the `metazone_id` field is empty.
-        #[diplomat::rust_link(icu::timezone::CustomTimeZone::metazone_id, StructField)]
-        #[diplomat::rust_link(icu::timezone::MetazoneId, Struct, compact)]
-        #[diplomat::attr(supports = accessors, getter)]
-        pub fn metazone_id(
-            &self,
-            write: &mut diplomat_runtime::DiplomatWriteable,
-        ) -> Result<(), ICU4XError> {
-            write.write_str(
-                self.0
-                    .metazone_id
-                    .ok_or(ICU4XError::TimeZoneMissingInputError)?
-                    .0
-                    .as_str(),
-            )?;
-            Ok(())
-        }
-
-        /// Sets the `zone_variant` field from a string.
-        ///
-        /// Errors if the string is not a valid zone variant.
-        #[diplomat::rust_link(icu::timezone::CustomTimeZone::zone_variant, StructField)]
-        #[diplomat::rust_link(icu::timezone::ZoneVariant, Struct, compact)]
-        #[diplomat::rust_link(icu::timezone::ZoneVariant::from_str, FnInStruct, hidden)]
-        pub fn try_set_zone_variant(&mut self, id: &DiplomatStr) -> Result<(), ICU4XError> {
-            self.0.zone_variant = Some(icu_timezone::ZoneVariant(
-                tinystr::TinyAsciiStr::from_bytes(id)
-                    .map_err(|_| ICU4XError::TimeZoneInvalidIdError)?,
-            ));
-            Ok(())
-        }
-
-        /// Clears the `zone_variant` field.
-        #[diplomat::rust_link(icu::timezone::CustomTimeZone::zone_variant, StructField)]
-        #[diplomat::rust_link(icu::timezone::ZoneVariant, Struct, compact)]
-        pub fn clear_zone_variant(&mut self) {
-            self.0.zone_variant.take();
-        }
-
-        /// Writes the value of the `zone_variant` field as a string.
-        ///
-        /// Errors if the `zone_variant` field is empty.
-        #[diplomat::rust_link(icu::timezone::CustomTimeZone::zone_variant, StructField)]
-        #[diplomat::rust_link(icu::timezone::ZoneVariant, Struct, compact)]
-        #[diplomat::attr(supports = accessors, getter)]
-        pub fn zone_variant(
-            &self,
-            write: &mut diplomat_runtime::DiplomatWriteable,
-        ) -> Result<(), ICU4XError> {
-            write.write_str(
-                self.0
-                    .zone_variant
-                    .ok_or(ICU4XError::TimeZoneMissingInputError)?
-                    .0
-                    .as_str(),
-            )?;
-            Ok(())
-        }
-
-        /// Sets the `zone_variant` field to "standard" time, which may or may
-        /// not correspond to a display name with "Standard" in its name.
-        #[diplomat::rust_link(icu::timezone::ZoneVariant::standard, FnInStruct)]
-        #[diplomat::rust_link(icu::timezone::CustomTimeZone::zone_variant, StructField, compact)]
-        pub fn set_standard_time(&mut self) {
-            self.0.zone_variant = Some(ZoneVariant::standard())
-        }
-
-        /// Sets the `zone_variant` field to "daylight" time, which may or may
-        /// not correspond to a display name with "Daylight" in its name.
-        #[diplomat::rust_link(icu::timezone::ZoneVariant::daylight, FnInStruct)]
-        #[diplomat::rust_link(icu::timezone::CustomTimeZone::zone_variant, StructField, compact)]
-        pub fn set_daylight_time(&mut self) {
-            self.0.zone_variant = Some(ZoneVariant::daylight())
-        }
-
-        /// Returns whether the `zone_variant` field is standard time.
-        ///
-        /// Errors if the `zone_variant` field is empty.
-        #[diplomat::rust_link(icu::timezone::ZoneVariant::standard, FnInStruct)]
-        #[diplomat::rust_link(icu::timezone::CustomTimeZone::zone_variant, StructField, compact)]
-        #[diplomat::attr(supports = accessors, getter)]
-        pub fn is_standard_time(&self) -> Result<bool, ICU4XError> {
-            Ok(self
-                .0
-                .zone_variant
-                .ok_or(ICU4XError::TimeZoneMissingInputError)?
-                == ZoneVariant::standard())
-        }
-
-        /// Returns whether the `zone_variant` field is daylight time.
-        ///
-        /// Errors if the `zone_variant` field is empty.
-        #[diplomat::rust_link(icu::timezone::ZoneVariant::daylight, FnInStruct)]
-        #[diplomat::rust_link(icu::timezone::CustomTimeZone::zone_variant, StructField, compact)]
-        #[diplomat::attr(supports = accessors, getter)]
-        pub fn is_daylight_time(&self) -> Result<bool, ICU4XError> {
-            Ok(self
-                .0
-                .zone_variant
-                .ok_or(ICU4XError::TimeZoneMissingInputError)?
-                == ZoneVariant::daylight())
-        }
-
-        /// Sets the metazone based on the time zone and the local timestamp.
-        #[diplomat::rust_link(icu::timezone::CustomTimeZone::maybe_calculate_metazone, FnInStruct)]
         #[diplomat::rust_link(
-            icu::timezone::MetazoneCalculator::compute_metazone_from_time_zone,
+            icu::time::zone::ZoneNameTimestamp::far_in_future,
             FnInStruct,
-            compact
+            hidden
+        )] // documented
+        #[diplomat::rust_link(icu::time::zone::ZoneNameTimestamp::far_in_past, FnInStruct, hidden)] // documented
+        pub fn at_date_time_iso(&self, date: &IsoDate, time: &Time) -> Box<Self> {
+            Box::new(Self {
+                zone_name_timestamp: Some(icu_time::zone::ZoneNameTimestamp::from_date_time_iso(
+                    icu_time::DateTime {
+                        date: date.0,
+                        time: time.0,
+                    },
+                )),
+                ..*self
+            })
+        }
+
+        #[diplomat::rust_link(icu::time::TimeZoneInfo::zone_name_timestamp, FnInStruct)]
+        #[diplomat::rust_link(
+            icu::time::zone::ZoneNameTimestamp::to_date_time_iso,
+            FnInStruct,
+            hidden
         )]
-        #[cfg(feature = "icu_timezone")]
-        pub fn maybe_calculate_metazone(
+        pub fn zone_name_date_time(&self) -> Option<IsoDateTime> {
+            let datetime = self.zone_name_timestamp?.to_date_time_iso();
+            Some(IsoDateTime {
+                date: Box::new(IsoDate(datetime.date)),
+                time: Box::new(Time(datetime.time)),
+            })
+        }
+
+        #[diplomat::rust_link(icu::time::TimeZoneInfo::with_variant, FnInStruct)]
+        pub fn with_variant(&self, time_variant: TimeZoneVariant) -> Box<Self> {
+            Box::new(Self {
+                variant: Some(time_variant.into()),
+                ..*self
+            })
+        }
+
+        /// Infers the zone variant.
+        ///
+        /// Requires the offset and local time to be set.
+        #[diplomat::rust_link(icu::time::TimeZoneInfo::infer_variant, FnInStruct)]
+        #[diplomat::rust_link(icu::time::zone::TimeZoneVariant, Enum, compact)]
+        pub fn infer_variant(
             &mut self,
-            metazone_calculator: &crate::metazone_calculator::ffi::ICU4XMetazoneCalculator,
-            local_datetime: &crate::datetime::ffi::ICU4XIsoDateTime,
-        ) {
-            self.0
-                .maybe_calculate_metazone(&metazone_calculator.0, &local_datetime.0);
+            offset_calculator: &VariantOffsetsCalculator,
+        ) -> Option<()> {
+            let info = self
+                .id
+                .with_offset(self.offset)
+                .with_zone_name_timestamp(self.zone_name_timestamp?)
+                .infer_variant(offset_calculator.0.as_borrowed());
+
+            self.id = info.id();
+            self.variant = Some(info.variant());
+            Some(())
+        }
+
+        #[diplomat::rust_link(icu::time::TimeZoneInfo::variant, FnInStruct)]
+        pub fn variant(&self) -> Option<TimeZoneVariant> {
+            self.variant.map(Into::into)
         }
     }
 }
 
-impl From<CustomTimeZone> for ffi::ICU4XCustomTimeZone {
-    fn from(other: CustomTimeZone) -> Self {
-        Self(other)
+impl From<icu_time::zone::UtcOffset> for TimeZoneInfo {
+    fn from(other: icu_time::zone::UtcOffset) -> Self {
+        Self {
+            id: icu_time::TimeZone::UNKNOWN,
+            offset: Some(other),
+            variant: None,
+            zone_name_timestamp: None,
+        }
     }
 }
 
-impl From<ffi::ICU4XCustomTimeZone> for CustomTimeZone {
-    fn from(other: ffi::ICU4XCustomTimeZone) -> Self {
-        other.0
+impl From<icu_time::TimeZoneInfo<icu_time::zone::models::Base>> for TimeZoneInfo {
+    fn from(other: icu_time::TimeZoneInfo<icu_time::zone::models::Base>) -> Self {
+        Self {
+            id: other.id(),
+            offset: other.offset(),
+            variant: None,
+            zone_name_timestamp: None,
+        }
+    }
+}
+
+impl From<icu_time::TimeZoneInfo<icu_time::zone::models::AtTime>> for TimeZoneInfo {
+    fn from(other: icu_time::TimeZoneInfo<icu_time::zone::models::AtTime>) -> Self {
+        Self {
+            id: other.id(),
+            offset: other.offset(),
+            variant: None,
+            zone_name_timestamp: Some(other.zone_name_timestamp()),
+        }
+    }
+}
+
+impl From<icu_time::TimeZoneInfo<icu_time::zone::models::Full>> for TimeZoneInfo {
+    fn from(other: icu_time::TimeZoneInfo<icu_time::zone::models::Full>) -> Self {
+        Self {
+            id: other.id(),
+            offset: other.offset(),
+            variant: Some(other.variant()),
+            zone_name_timestamp: Some(other.zone_name_timestamp()),
+        }
     }
 }
