@@ -2,312 +2,501 @@
 // called LICENSE at the top level of the ICU4X source tree
 // (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
 
-use icu_datetime::time_zone::{FallbackFormat, TimeZoneFormatterOptions};
+// @generated by tools/make/codegen/src/capi_datetime.rs
 
-macro_rules! call_method {
-    ($self:ident, $compiled:ident, $unstable:ident, $provider:expr) => {
-        match &$provider.0 {
-            $crate::provider::ICU4XDataProviderInner::Destroyed => Err(
-                icu_provider::DataError::custom("This provider has been destroyed"),
-            )?,
-            $crate::provider::ICU4XDataProviderInner::Empty => $self
-                .0
-                .$unstable(&icu_provider_adapters::empty::EmptyDataProvider::new()),
-            #[cfg(feature = "buffer_provider")]
-            $crate::provider::ICU4XDataProviderInner::Buffer(buffer_provider) => $self.0.$unstable(
-                &icu_provider::AsDeserializingBufferProvider::as_deserializing(buffer_provider),
-            ),
-            #[cfg(feature = "compiled_data")]
-            $crate::provider::ICU4XDataProviderInner::Compiled => $self.0.$compiled(),
-        }
-    };
-}
-
+#[rustfmt::skip]
 #[diplomat::bridge]
+#[diplomat::abi_rename = "icu4x_{0}_mv1"]
+#[diplomat::attr(auto, namespace = "icu4x")]
 pub mod ffi {
-    use crate::errors::ffi::ICU4XError;
-    use crate::locale::ffi::ICU4XLocale;
-    use crate::provider::ffi::ICU4XDataProvider;
-    use crate::timezone::ffi::ICU4XCustomTimeZone;
     use alloc::boxed::Box;
-    use icu_datetime::time_zone::FallbackFormat;
-    use icu_datetime::time_zone::IsoFormat;
-    use icu_datetime::time_zone::IsoMinutes;
-    use icu_datetime::time_zone::IsoSeconds;
-    use icu_datetime::time_zone::TimeZoneFormatter;
-    use writeable::Writeable;
+    use writeable::TryWriteable;
+
+    #[allow(unused_imports)]
+    use crate::datetime_helpers::{self, map_or_default};
+
+    #[allow(unused_imports)]
+    use crate::unstable::{
+        date_formatter::ffi::{DateFormatter, DateFormatterGregorian},
+        date_time_formatter::ffi::{DateTimeFormatter, DateTimeFormatterGregorian},
+        date::ffi::IsoDate,
+        datetime_options::ffi::{DateTimeAlignment, DateTimeLength, TimePrecision},
+        errors::ffi::DateTimeFormatterLoadError,
+        errors::ffi::DateTimeWriteError,
+        locale_core::ffi::Locale,
+        time_formatter::ffi::TimeFormatter,
+        time::ffi::Time,
+        timezone::ffi::TimeZoneInfo,
+    };
+
+    #[cfg(feature = "buffer_provider")]
+    use crate::unstable::provider::ffi::DataProvider;
 
     #[diplomat::opaque]
-    /// An ICU4X TimeZoneFormatter object capable of formatting an [`ICU4XCustomTimeZone`] type (and others) as a string
-    #[diplomat::rust_link(icu::datetime::time_zone::TimeZoneFormatter, Struct)]
-    #[diplomat::rust_link(icu::datetime::FormattedTimeZone, Struct, hidden)]
-    pub struct ICU4XTimeZoneFormatter(pub TimeZoneFormatter);
+    #[diplomat::rust_link(icu::datetime::NoCalendarFormatter, Typedef)]
+    pub struct TimeZoneFormatter(
+        pub  icu_datetime::FixedCalendarDateTimeFormatter<
+            (),
+            icu_datetime::fieldsets::enums::ZoneFieldSet,
+        >,
+    );
 
-    #[diplomat::enum_convert(IsoFormat, needs_wildcard)]
-    #[diplomat::rust_link(icu::datetime::time_zone::IsoFormat, Enum)]
-    pub enum ICU4XIsoTimeZoneFormat {
-        Basic,
-        Extended,
-        UtcBasic,
-        UtcExtended,
-    }
-
-    #[diplomat::enum_convert(IsoMinutes, needs_wildcard)]
-    #[diplomat::rust_link(icu::datetime::time_zone::IsoMinutes, Enum)]
-    pub enum ICU4XIsoTimeZoneMinuteDisplay {
-        Required,
-        Optional,
-    }
-
-    #[diplomat::enum_convert(IsoSeconds, needs_wildcard)]
-    #[diplomat::rust_link(icu::datetime::time_zone::IsoSeconds, Enum)]
-    pub enum ICU4XIsoTimeZoneSecondDisplay {
-        Optional,
-        Never,
-    }
-
-    pub struct ICU4XIsoTimeZoneOptions {
-        pub format: ICU4XIsoTimeZoneFormat,
-        pub minutes: ICU4XIsoTimeZoneMinuteDisplay,
-        pub seconds: ICU4XIsoTimeZoneSecondDisplay,
-    }
-
-    impl ICU4XTimeZoneFormatter {
-        /// Creates a new [`ICU4XTimeZoneFormatter`] from locale data.
+    impl TimeZoneFormatter {
+        /// Creates a zoned formatter based on a non-zoned formatter.
         ///
-        /// Uses localized GMT as the fallback format.
-        #[diplomat::rust_link(icu::datetime::time_zone::TimeZoneFormatter::try_new, FnInStruct)]
-        #[diplomat::rust_link(icu::datetime::time_zone::FallbackFormat, Enum, compact)]
-        #[diplomat::rust_link(icu::datetime::time_zone::TimeZoneFormatterOptions, Struct, hidden)]
-        #[diplomat::attr(all(supports = constructors, supports = fallible_constructors, supports = named_constructors), named_constructor = "with_localized_gmt_fallback")]
-        pub fn create_with_localized_gmt_fallback(
-            provider: &ICU4XDataProvider,
-            locale: &ICU4XLocale,
-        ) -> Result<Box<ICU4XTimeZoneFormatter>, ICU4XError> {
-            let locale = locale.to_datalocale();
-
-            Ok(Box::new(ICU4XTimeZoneFormatter(call_constructor!(
-                TimeZoneFormatter::try_new,
-                TimeZoneFormatter::try_new_with_any_provider,
-                TimeZoneFormatter::try_new_with_buffer_provider,
-                provider,
-                &locale,
-                FallbackFormat::LocalizedGmt.into(),
-            )?)))
+        /// Caution: The locale provided here must match the locale used to construct the non-zoned formatter,
+        /// or else unexpected behavior may occur!
+        #[diplomat::attr(all(supports = fallible_constructors, supports = named_constructors), named_constructor = "specific_long")]
+        #[diplomat::rust_link(icu::datetime::fieldsets::zone::SpecificLong, Struct)]
+        #[diplomat::rust_link(icu::datetime::fieldsets::Combo, Struct, hidden)]
+        #[cfg(feature = "compiled_data")]
+        pub fn create_specific_long(
+            locale: &Locale,
+        ) -> Result<Box<Self>, DateTimeFormatterLoadError> {
+            let zone = icu_datetime::fieldsets::zone::SpecificLong;
+            let prefs = (&locale.0).into();
+            let options = zone;
+            Ok(Box::new(Self(
+                icu_datetime
+                    ::FixedCalendarDateTimeFormatter
+                    ::try_new(
+                        prefs,
+                        options
+                    )?
+                .cast_into_fset(),
+            )))
         }
-
-        /// Creates a new [`ICU4XTimeZoneFormatter`] from locale data.
+        
+        /// Creates a zoned formatter based on a non-zoned formatter.
         ///
-        /// Uses ISO-8601 as the fallback format.
-        #[diplomat::rust_link(icu::datetime::time_zone::TimeZoneFormatter::try_new, FnInStruct)]
-        #[diplomat::rust_link(icu::datetime::time_zone::FallbackFormat, Enum, compact)]
-        #[diplomat::rust_link(icu::datetime::time_zone::TimeZoneFormatterOptions, Struct, hidden)]
-        #[diplomat::attr(all(supports = constructors, supports = fallible_constructors, supports = named_constructors), named_constructor = "with_iso_8601_fallback")]
-        pub fn create_with_iso_8601_fallback(
-            provider: &ICU4XDataProvider,
-            locale: &ICU4XLocale,
-            options: ICU4XIsoTimeZoneOptions,
-        ) -> Result<Box<ICU4XTimeZoneFormatter>, ICU4XError> {
-            let locale = locale.to_datalocale();
-
-            Ok(Box::new(ICU4XTimeZoneFormatter(call_constructor!(
-                TimeZoneFormatter::try_new,
-                TimeZoneFormatter::try_new_with_any_provider,
-                TimeZoneFormatter::try_new_with_buffer_provider,
-                provider,
-                &locale,
-                options.into(),
-            )?)))
+        /// Caution: The locale provided here must match the locale used to construct the non-zoned formatter,
+        /// or else unexpected behavior may occur!
+        #[diplomat::attr(all(supports = fallible_constructors, supports = named_constructors), named_constructor = "specific_long_with_provider")]
+        #[diplomat::rust_link(icu::datetime::fieldsets::zone::SpecificLong, Struct)]
+        #[diplomat::rust_link(icu::datetime::fieldsets::Combo, Struct, hidden)]
+        #[cfg(feature = "buffer_provider")]
+        pub fn create_specific_long_with_provider(
+            provider: &DataProvider,
+            locale: &Locale,
+        ) -> Result<Box<Self>, DateTimeFormatterLoadError> {
+            let provider = provider.get()?;
+            let zone = icu_datetime::fieldsets::zone::SpecificLong;
+            let prefs = (&locale.0).into();
+            let options = zone;
+            Ok(Box::new(Self(
+                icu_datetime
+                    ::FixedCalendarDateTimeFormatter
+                    ::try_new_with_buffer_provider(
+                        provider,
+                        prefs,
+                        options
+                    )?
+                .cast_into_fset(),
+            )))
         }
-
-        /// Loads generic non-location long format. Example: "Pacific Time"
-        #[diplomat::rust_link(
-            icu::datetime::time_zone::TimeZoneFormatter::include_generic_non_location_long,
-            FnInStruct
-        )]
-        #[diplomat::rust_link(
-            icu::datetime::time_zone::TimeZoneFormatter::load_generic_non_location_long,
-            FnInStruct,
-            hidden
-        )]
-        pub fn load_generic_non_location_long(
-            &mut self,
-            provider: &ICU4XDataProvider,
-        ) -> Result<(), ICU4XError> {
-            call_method!(
-                self,
-                include_generic_non_location_long,
-                load_generic_non_location_long,
-                provider
-            )?;
-            Ok(())
+        
+        /// Creates a zoned formatter based on a non-zoned formatter.
+        ///
+        /// Caution: The locale provided here must match the locale used to construct the non-zoned formatter,
+        /// or else unexpected behavior may occur!
+        #[diplomat::attr(all(supports = fallible_constructors, supports = named_constructors), named_constructor = "specific_short")]
+        #[diplomat::rust_link(icu::datetime::fieldsets::zone::SpecificShort, Struct)]
+        #[diplomat::rust_link(icu::datetime::fieldsets::Combo, Struct, hidden)]
+        #[cfg(feature = "compiled_data")]
+        pub fn create_specific_short(
+            locale: &Locale,
+        ) -> Result<Box<Self>, DateTimeFormatterLoadError> {
+            let zone = icu_datetime::fieldsets::zone::SpecificShort;
+            let prefs = (&locale.0).into();
+            let options = zone;
+            Ok(Box::new(Self(
+                icu_datetime
+                    ::FixedCalendarDateTimeFormatter
+                    ::try_new(
+                        prefs,
+                        options
+                    )?
+                .cast_into_fset(),
+            )))
         }
-
-        /// Loads generic non-location short format. Example: "PT"
-        #[diplomat::rust_link(
-            icu::datetime::time_zone::TimeZoneFormatter::include_generic_non_location_short,
-            FnInStruct
-        )]
-        #[diplomat::rust_link(
-            icu::datetime::time_zone::TimeZoneFormatter::load_generic_non_location_short,
-            FnInStruct,
-            hidden
-        )]
-        pub fn load_generic_non_location_short(
-            &mut self,
-            provider: &ICU4XDataProvider,
-        ) -> Result<(), ICU4XError> {
-            call_method!(
-                self,
-                include_generic_non_location_short,
-                load_generic_non_location_short,
-                provider
-            )?;
-            Ok(())
+        
+        /// Creates a zoned formatter based on a non-zoned formatter.
+        ///
+        /// Caution: The locale provided here must match the locale used to construct the non-zoned formatter,
+        /// or else unexpected behavior may occur!
+        #[diplomat::attr(all(supports = fallible_constructors, supports = named_constructors), named_constructor = "specific_short_with_provider")]
+        #[diplomat::rust_link(icu::datetime::fieldsets::zone::SpecificShort, Struct)]
+        #[diplomat::rust_link(icu::datetime::fieldsets::Combo, Struct, hidden)]
+        #[cfg(feature = "buffer_provider")]
+        pub fn create_specific_short_with_provider(
+            provider: &DataProvider,
+            locale: &Locale,
+        ) -> Result<Box<Self>, DateTimeFormatterLoadError> {
+            let provider = provider.get()?;
+            let zone = icu_datetime::fieldsets::zone::SpecificShort;
+            let prefs = (&locale.0).into();
+            let options = zone;
+            Ok(Box::new(Self(
+                icu_datetime
+                    ::FixedCalendarDateTimeFormatter
+                    ::try_new_with_buffer_provider(
+                        provider,
+                        prefs,
+                        options
+                    )?
+                .cast_into_fset(),
+            )))
         }
-
-        /// Loads specific non-location long format. Example: "Pacific Standard Time"
-        #[diplomat::rust_link(
-            icu::datetime::time_zone::TimeZoneFormatter::include_specific_non_location_long,
-            FnInStruct
-        )]
-        #[diplomat::rust_link(
-            icu::datetime::time_zone::TimeZoneFormatter::load_specific_non_location_long,
-            FnInStruct,
-            hidden
-        )]
-        pub fn load_specific_non_location_long(
-            &mut self,
-            provider: &ICU4XDataProvider,
-        ) -> Result<(), ICU4XError> {
-            call_method!(
-                self,
-                include_specific_non_location_long,
-                load_specific_non_location_long,
-                provider
-            )?;
-            Ok(())
+        
+        /// Creates a zoned formatter based on a non-zoned formatter.
+        ///
+        /// Caution: The locale provided here must match the locale used to construct the non-zoned formatter,
+        /// or else unexpected behavior may occur!
+        #[diplomat::attr(all(supports = fallible_constructors, supports = named_constructors), named_constructor = "localized_offset_long")]
+        #[diplomat::rust_link(icu::datetime::fieldsets::zone::LocalizedOffsetLong, Struct)]
+        #[diplomat::rust_link(icu::datetime::fieldsets::Combo, Struct, hidden)]
+        #[cfg(feature = "compiled_data")]
+        pub fn create_localized_offset_long(
+            locale: &Locale,
+        ) -> Result<Box<Self>, DateTimeFormatterLoadError> {
+            let zone = icu_datetime::fieldsets::zone::LocalizedOffsetLong;
+            let prefs = (&locale.0).into();
+            let options = zone;
+            Ok(Box::new(Self(
+                icu_datetime
+                    ::FixedCalendarDateTimeFormatter
+                    ::try_new(
+                        prefs,
+                        options
+                    )?
+                .cast_into_fset(),
+            )))
         }
-
-        /// Loads specific non-location short format. Example: "PST"
-        #[diplomat::rust_link(
-            icu::datetime::time_zone::TimeZoneFormatter::include_specific_non_location_short,
-            FnInStruct
-        )]
-        #[diplomat::rust_link(
-            icu::datetime::time_zone::TimeZoneFormatter::load_specific_non_location_short,
-            FnInStruct,
-            hidden
-        )]
-        pub fn load_specific_non_location_short(
-            &mut self,
-            provider: &ICU4XDataProvider,
-        ) -> Result<(), ICU4XError> {
-            call_method!(
-                self,
-                include_specific_non_location_short,
-                load_specific_non_location_short,
-                provider
-            )?;
-            Ok(())
+        
+        /// Creates a zoned formatter based on a non-zoned formatter.
+        ///
+        /// Caution: The locale provided here must match the locale used to construct the non-zoned formatter,
+        /// or else unexpected behavior may occur!
+        #[diplomat::attr(all(supports = fallible_constructors, supports = named_constructors), named_constructor = "localized_offset_long_with_provider")]
+        #[diplomat::rust_link(icu::datetime::fieldsets::zone::LocalizedOffsetLong, Struct)]
+        #[diplomat::rust_link(icu::datetime::fieldsets::Combo, Struct, hidden)]
+        #[cfg(feature = "buffer_provider")]
+        pub fn create_localized_offset_long_with_provider(
+            provider: &DataProvider,
+            locale: &Locale,
+        ) -> Result<Box<Self>, DateTimeFormatterLoadError> {
+            let provider = provider.get()?;
+            let zone = icu_datetime::fieldsets::zone::LocalizedOffsetLong;
+            let prefs = (&locale.0).into();
+            let options = zone;
+            Ok(Box::new(Self(
+                icu_datetime
+                    ::FixedCalendarDateTimeFormatter
+                    ::try_new_with_buffer_provider(
+                        provider,
+                        prefs,
+                        options
+                    )?
+                .cast_into_fset(),
+            )))
         }
-
-        /// Loads generic location format. Example: "Los Angeles Time"
-        #[diplomat::rust_link(
-            icu::datetime::time_zone::TimeZoneFormatter::include_generic_location_format,
-            FnInStruct
-        )]
-        #[diplomat::rust_link(
-            icu::datetime::time_zone::TimeZoneFormatter::load_generic_location_format,
-            FnInStruct,
-            hidden
-        )]
-        pub fn load_generic_location_format(
-            &mut self,
-            provider: &ICU4XDataProvider,
-        ) -> Result<(), ICU4XError> {
-            call_method!(
-                self,
-                include_generic_location_format,
-                load_generic_location_format,
-                provider
-            )?;
-            Ok(())
+        
+        /// Creates a zoned formatter based on a non-zoned formatter.
+        ///
+        /// Caution: The locale provided here must match the locale used to construct the non-zoned formatter,
+        /// or else unexpected behavior may occur!
+        #[diplomat::attr(all(supports = fallible_constructors, supports = named_constructors), named_constructor = "localized_offset_short")]
+        #[diplomat::rust_link(icu::datetime::fieldsets::zone::LocalizedOffsetShort, Struct)]
+        #[diplomat::rust_link(icu::datetime::fieldsets::Combo, Struct, hidden)]
+        #[cfg(feature = "compiled_data")]
+        pub fn create_localized_offset_short(
+            locale: &Locale,
+        ) -> Result<Box<Self>, DateTimeFormatterLoadError> {
+            let zone = icu_datetime::fieldsets::zone::LocalizedOffsetShort;
+            let prefs = (&locale.0).into();
+            let options = zone;
+            Ok(Box::new(Self(
+                icu_datetime
+                    ::FixedCalendarDateTimeFormatter
+                    ::try_new(
+                        prefs,
+                        options
+                    )?
+                .cast_into_fset(),
+            )))
         }
-
-        /// Loads localized GMT format. Example: "GMT-07:00"
-        #[diplomat::rust_link(
-            icu::datetime::time_zone::TimeZoneFormatter::include_localized_gmt_format,
-            FnInStruct
-        )]
-        #[diplomat::rust_link(
-            icu::datetime::time_zone::TimeZoneFormatter::load_localized_gmt_format,
-            FnInStruct,
-            hidden
-        )]
-        pub fn include_localized_gmt_format(&mut self) -> Result<(), ICU4XError> {
-            self.0.include_localized_gmt_format()?;
-            Ok(())
+        
+        /// Creates a zoned formatter based on a non-zoned formatter.
+        ///
+        /// Caution: The locale provided here must match the locale used to construct the non-zoned formatter,
+        /// or else unexpected behavior may occur!
+        #[diplomat::attr(all(supports = fallible_constructors, supports = named_constructors), named_constructor = "localized_offset_short_with_provider")]
+        #[diplomat::rust_link(icu::datetime::fieldsets::zone::LocalizedOffsetShort, Struct)]
+        #[diplomat::rust_link(icu::datetime::fieldsets::Combo, Struct, hidden)]
+        #[cfg(feature = "buffer_provider")]
+        pub fn create_localized_offset_short_with_provider(
+            provider: &DataProvider,
+            locale: &Locale,
+        ) -> Result<Box<Self>, DateTimeFormatterLoadError> {
+            let provider = provider.get()?;
+            let zone = icu_datetime::fieldsets::zone::LocalizedOffsetShort;
+            let prefs = (&locale.0).into();
+            let options = zone;
+            Ok(Box::new(Self(
+                icu_datetime
+                    ::FixedCalendarDateTimeFormatter
+                    ::try_new_with_buffer_provider(
+                        provider,
+                        prefs,
+                        options
+                    )?
+                .cast_into_fset(),
+            )))
         }
-
-        /// Loads ISO-8601 format. Example: "-07:00"
-        #[diplomat::rust_link(
-            icu::datetime::time_zone::TimeZoneFormatter::include_iso_8601_format,
-            FnInStruct
-        )]
-        #[diplomat::rust_link(
-            icu::datetime::time_zone::TimeZoneFormatter::load_iso_8601_format,
-            FnInStruct,
-            hidden
-        )]
-        pub fn load_iso_8601_format(
-            &mut self,
-            options: ICU4XIsoTimeZoneOptions,
-        ) -> Result<(), ICU4XError> {
-            self.0.include_iso_8601_format(
-                options.format.into(),
-                options.minutes.into(),
-                options.seconds.into(),
-            )?;
-            Ok(())
+        
+        /// Creates a zoned formatter based on a non-zoned formatter.
+        ///
+        /// Caution: The locale provided here must match the locale used to construct the non-zoned formatter,
+        /// or else unexpected behavior may occur!
+        #[diplomat::attr(all(supports = fallible_constructors, supports = named_constructors), named_constructor = "generic_long")]
+        #[diplomat::rust_link(icu::datetime::fieldsets::zone::GenericLong, Struct)]
+        #[diplomat::rust_link(icu::datetime::fieldsets::Combo, Struct, hidden)]
+        #[cfg(feature = "compiled_data")]
+        pub fn create_generic_long(
+            locale: &Locale,
+        ) -> Result<Box<Self>, DateTimeFormatterLoadError> {
+            let zone = icu_datetime::fieldsets::zone::GenericLong;
+            let prefs = (&locale.0).into();
+            let options = zone;
+            Ok(Box::new(Self(
+                icu_datetime
+                    ::FixedCalendarDateTimeFormatter
+                    ::try_new(
+                        prefs,
+                        options
+                    )?
+                .cast_into_fset(),
+            )))
         }
-
-        /// Formats a [`ICU4XCustomTimeZone`] to a string.
-        #[diplomat::rust_link(icu::datetime::time_zone::TimeZoneFormatter::format, FnInStruct)]
-        #[diplomat::rust_link(
-            icu::datetime::time_zone::TimeZoneFormatter::format_to_string,
-            FnInStruct
-        )]
-        #[diplomat::rust_link(icu::datetime::FormattedTimeZone::write_to, FnInStruct, hidden)]
-        pub fn format_custom_time_zone(
+        
+        /// Creates a zoned formatter based on a non-zoned formatter.
+        ///
+        /// Caution: The locale provided here must match the locale used to construct the non-zoned formatter,
+        /// or else unexpected behavior may occur!
+        #[diplomat::attr(all(supports = fallible_constructors, supports = named_constructors), named_constructor = "generic_long_with_provider")]
+        #[diplomat::rust_link(icu::datetime::fieldsets::zone::GenericLong, Struct)]
+        #[diplomat::rust_link(icu::datetime::fieldsets::Combo, Struct, hidden)]
+        #[cfg(feature = "buffer_provider")]
+        pub fn create_generic_long_with_provider(
+            provider: &DataProvider,
+            locale: &Locale,
+        ) -> Result<Box<Self>, DateTimeFormatterLoadError> {
+            let provider = provider.get()?;
+            let zone = icu_datetime::fieldsets::zone::GenericLong;
+            let prefs = (&locale.0).into();
+            let options = zone;
+            Ok(Box::new(Self(
+                icu_datetime
+                    ::FixedCalendarDateTimeFormatter
+                    ::try_new_with_buffer_provider(
+                        provider,
+                        prefs,
+                        options
+                    )?
+                .cast_into_fset(),
+            )))
+        }
+        
+        /// Creates a zoned formatter based on a non-zoned formatter.
+        ///
+        /// Caution: The locale provided here must match the locale used to construct the non-zoned formatter,
+        /// or else unexpected behavior may occur!
+        #[diplomat::attr(all(supports = fallible_constructors, supports = named_constructors), named_constructor = "generic_short")]
+        #[diplomat::rust_link(icu::datetime::fieldsets::zone::GenericShort, Struct)]
+        #[diplomat::rust_link(icu::datetime::fieldsets::Combo, Struct, hidden)]
+        #[diplomat::demo(default_constructor)]
+        #[cfg(feature = "compiled_data")]
+        pub fn create_generic_short(
+            locale: &Locale,
+        ) -> Result<Box<Self>, DateTimeFormatterLoadError> {
+            let zone = icu_datetime::fieldsets::zone::GenericShort;
+            let prefs = (&locale.0).into();
+            let options = zone;
+            Ok(Box::new(Self(
+                icu_datetime
+                    ::FixedCalendarDateTimeFormatter
+                    ::try_new(
+                        prefs,
+                        options
+                    )?
+                .cast_into_fset(),
+            )))
+        }
+        
+        /// Creates a zoned formatter based on a non-zoned formatter.
+        ///
+        /// Caution: The locale provided here must match the locale used to construct the non-zoned formatter,
+        /// or else unexpected behavior may occur!
+        #[diplomat::attr(all(supports = fallible_constructors, supports = named_constructors), named_constructor = "generic_short_with_provider")]
+        #[diplomat::rust_link(icu::datetime::fieldsets::zone::GenericShort, Struct)]
+        #[diplomat::rust_link(icu::datetime::fieldsets::Combo, Struct, hidden)]
+        #[cfg(feature = "buffer_provider")]
+        pub fn create_generic_short_with_provider(
+            provider: &DataProvider,
+            locale: &Locale,
+        ) -> Result<Box<Self>, DateTimeFormatterLoadError> {
+            let provider = provider.get()?;
+            let zone = icu_datetime::fieldsets::zone::GenericShort;
+            let prefs = (&locale.0).into();
+            let options = zone;
+            Ok(Box::new(Self(
+                icu_datetime
+                    ::FixedCalendarDateTimeFormatter
+                    ::try_new_with_buffer_provider(
+                        provider,
+                        prefs,
+                        options
+                    )?
+                .cast_into_fset(),
+            )))
+        }
+        
+        /// Creates a zoned formatter based on a non-zoned formatter.
+        ///
+        /// Caution: The locale provided here must match the locale used to construct the non-zoned formatter,
+        /// or else unexpected behavior may occur!
+        #[diplomat::attr(all(supports = fallible_constructors, supports = named_constructors), named_constructor = "location")]
+        #[diplomat::rust_link(icu::datetime::fieldsets::zone::Location, Struct)]
+        #[diplomat::rust_link(icu::datetime::fieldsets::Combo, Struct, hidden)]
+        #[cfg(feature = "compiled_data")]
+        pub fn create_location(
+            locale: &Locale,
+        ) -> Result<Box<Self>, DateTimeFormatterLoadError> {
+            let zone = icu_datetime::fieldsets::zone::Location;
+            let prefs = (&locale.0).into();
+            let options = zone;
+            Ok(Box::new(Self(
+                icu_datetime
+                    ::FixedCalendarDateTimeFormatter
+                    ::try_new(
+                        prefs,
+                        options
+                    )?
+                .cast_into_fset(),
+            )))
+        }
+        
+        /// Creates a zoned formatter based on a non-zoned formatter.
+        ///
+        /// Caution: The locale provided here must match the locale used to construct the non-zoned formatter,
+        /// or else unexpected behavior may occur!
+        #[diplomat::attr(all(supports = fallible_constructors, supports = named_constructors), named_constructor = "location_with_provider")]
+        #[diplomat::rust_link(icu::datetime::fieldsets::zone::Location, Struct)]
+        #[diplomat::rust_link(icu::datetime::fieldsets::Combo, Struct, hidden)]
+        #[cfg(feature = "buffer_provider")]
+        pub fn create_location_with_provider(
+            provider: &DataProvider,
+            locale: &Locale,
+        ) -> Result<Box<Self>, DateTimeFormatterLoadError> {
+            let provider = provider.get()?;
+            let zone = icu_datetime::fieldsets::zone::Location;
+            let prefs = (&locale.0).into();
+            let options = zone;
+            Ok(Box::new(Self(
+                icu_datetime
+                    ::FixedCalendarDateTimeFormatter
+                    ::try_new_with_buffer_provider(
+                        provider,
+                        prefs,
+                        options
+                    )?
+                .cast_into_fset(),
+            )))
+        }
+        
+        /// Creates a zoned formatter based on a non-zoned formatter.
+        ///
+        /// Caution: The locale provided here must match the locale used to construct the non-zoned formatter,
+        /// or else unexpected behavior may occur!
+        #[diplomat::attr(all(supports = fallible_constructors, supports = named_constructors), named_constructor = "exemplar_city")]
+        #[diplomat::rust_link(icu::datetime::fieldsets::zone::ExemplarCity, Struct)]
+        #[diplomat::rust_link(icu::datetime::fieldsets::Combo, Struct, hidden)]
+        #[cfg(feature = "compiled_data")]
+        pub fn create_exemplar_city(
+            locale: &Locale,
+        ) -> Result<Box<Self>, DateTimeFormatterLoadError> {
+            let zone = icu_datetime::fieldsets::zone::ExemplarCity;
+            let prefs = (&locale.0).into();
+            let options = zone;
+            Ok(Box::new(Self(
+                icu_datetime
+                    ::FixedCalendarDateTimeFormatter
+                    ::try_new(
+                        prefs,
+                        options
+                    )?
+                .cast_into_fset(),
+            )))
+        }
+        
+        /// Creates a zoned formatter based on a non-zoned formatter.
+        ///
+        /// Caution: The locale provided here must match the locale used to construct the non-zoned formatter,
+        /// or else unexpected behavior may occur!
+        #[diplomat::attr(all(supports = fallible_constructors, supports = named_constructors), named_constructor = "exemplar_city_with_provider")]
+        #[diplomat::rust_link(icu::datetime::fieldsets::zone::ExemplarCity, Struct)]
+        #[diplomat::rust_link(icu::datetime::fieldsets::Combo, Struct, hidden)]
+        #[cfg(feature = "buffer_provider")]
+        pub fn create_exemplar_city_with_provider(
+            provider: &DataProvider,
+            locale: &Locale,
+        ) -> Result<Box<Self>, DateTimeFormatterLoadError> {
+            let provider = provider.get()?;
+            let zone = icu_datetime::fieldsets::zone::ExemplarCity;
+            let prefs = (&locale.0).into();
+            let options = zone;
+            Ok(Box::new(Self(
+                icu_datetime
+                    ::FixedCalendarDateTimeFormatter
+                    ::try_new_with_buffer_provider(
+                        provider,
+                        prefs,
+                        options
+                    )?
+                .cast_into_fset(),
+            )))
+        }
+        
+        #[diplomat::rust_link(icu::datetime::FixedCalendarDateTimeFormatter::format, FnInStruct)]
+        #[diplomat::rust_link(icu::datetime::FormattedDateTime, Struct, hidden)]
+        #[diplomat::rust_link(icu::datetime::FormattedDateTime::to_string, FnInStruct, hidden)]
+        pub fn format(
             &self,
-            value: &ICU4XCustomTimeZone,
-            write: &mut diplomat_runtime::DiplomatWriteable,
-        ) -> Result<(), ICU4XError> {
-            self.0.format(&value.0).write_to(write)?;
-            Ok(())
-        }
-
-        /// Formats a [`ICU4XCustomTimeZone`] to a string, performing no fallback
-        #[diplomat::rust_link(icu::datetime::FormattedTimeZone::write_no_fallback, FnInStruct)]
-        pub fn format_custom_time_zone_no_fallback(
-            &self,
-            value: &ICU4XCustomTimeZone,
-            write: &mut diplomat_runtime::DiplomatWriteable,
-        ) -> Result<(), ICU4XError> {
-            self.0.format(&value.0).write_no_fallback(write)??;
+            zone: &TimeZoneInfo,
+            write: &mut diplomat_runtime::DiplomatWrite,
+        ) -> Result<(), DateTimeWriteError> {
+            let mut input = icu_datetime::unchecked::DateTimeInputUnchecked::default();
+            input.set_time_zone_id(zone.id);
+            if let Some(offset) = zone.offset {
+                input.set_time_zone_utc_offset(offset);
+            }
+            if let Some(zone_name_timestamp) = zone.zone_name_timestamp {
+                input.set_time_zone_name_timestamp(zone_name_timestamp);
+            }
+            else {
+                input.set_time_zone_name_timestamp(icu_time::zone::ZoneNameTimestamp::far_in_future())
+            }
+            if let Some(variant) = zone.variant {
+                input.set_time_zone_variant(variant);
+            }
+            let _infallible = self
+                .0
+                .format_unchecked(input)
+                .try_write_to(write)
+                .ok()
+                .transpose()?;
             Ok(())
         }
     }
-}
-
-impl From<ffi::ICU4XIsoTimeZoneOptions> for TimeZoneFormatterOptions {
-    fn from(other: ffi::ICU4XIsoTimeZoneOptions) -> Self {
-        FallbackFormat::Iso8601(
-            other.format.into(),
-            other.minutes.into(),
-            other.seconds.into(),
-        )
-        .into()
-    }
+    
 }
