@@ -1,2 +1,2 @@
 // @generated
-include ! ("macros.rs") ; macro_rules ! impl_data_provider { ($ provider : ty) => { make_provider ! ($ provider) ; impl_segmenter_dictionary_w_auto_v1 ! ($ provider) ; impl_segmenter_dictionary_wl_ext_v1 ! ($ provider) ; impl_segmenter_grapheme_v1 ! ($ provider) ; impl_segmenter_line_v1 ! ($ provider) ; impl_segmenter_lstm_wl_auto_v1 ! ($ provider) ; impl_segmenter_sentence_v1 ! ($ provider) ; impl_segmenter_word_v1 ! ($ provider) ; } ; } # [allow (unused_macros)] macro_rules ! impl_any_provider { ($ provider : ty) => { # [clippy :: msrv = "1.67"] impl icu_provider :: AnyProvider for $ provider { fn load_any (& self , key : icu_provider :: DataKey , req : icu_provider :: DataRequest) -> Result < icu_provider :: AnyResponse , icu_provider :: DataError > { match key . hashed () { h if h == < icu::segmenter :: provider :: DictionaryForWordOnlyAutoV1Marker as icu_provider :: KeyedDataMarker > :: KEY . hashed () => icu_provider :: DataProvider :: < icu::segmenter :: provider :: DictionaryForWordOnlyAutoV1Marker > :: load (self , req) . map (icu_provider :: DataResponse :: wrap_into_any_response) , h if h == < icu::segmenter :: provider :: DictionaryForWordLineExtendedV1Marker as icu_provider :: KeyedDataMarker > :: KEY . hashed () => icu_provider :: DataProvider :: < icu::segmenter :: provider :: DictionaryForWordLineExtendedV1Marker > :: load (self , req) . map (icu_provider :: DataResponse :: wrap_into_any_response) , h if h == < icu::segmenter :: provider :: GraphemeClusterBreakDataV1Marker as icu_provider :: KeyedDataMarker > :: KEY . hashed () => icu_provider :: DataProvider :: < icu::segmenter :: provider :: GraphemeClusterBreakDataV1Marker > :: load (self , req) . map (icu_provider :: DataResponse :: wrap_into_any_response) , h if h == < icu::segmenter :: provider :: LineBreakDataV1Marker as icu_provider :: KeyedDataMarker > :: KEY . hashed () => icu_provider :: DataProvider :: < icu::segmenter :: provider :: LineBreakDataV1Marker > :: load (self , req) . map (icu_provider :: DataResponse :: wrap_into_any_response) , h if h == < icu::segmenter :: provider :: LstmForWordLineAutoV1Marker as icu_provider :: KeyedDataMarker > :: KEY . hashed () => icu_provider :: DataProvider :: < icu::segmenter :: provider :: LstmForWordLineAutoV1Marker > :: load (self , req) . map (icu_provider :: DataResponse :: wrap_into_any_response) , h if h == < icu::segmenter :: provider :: SentenceBreakDataV1Marker as icu_provider :: KeyedDataMarker > :: KEY . hashed () => icu_provider :: DataProvider :: < icu::segmenter :: provider :: SentenceBreakDataV1Marker > :: load (self , req) . map (icu_provider :: DataResponse :: wrap_into_any_response) , h if h == < icu::segmenter :: provider :: WordBreakDataV1Marker as icu_provider :: KeyedDataMarker > :: KEY . hashed () => icu_provider :: DataProvider :: < icu::segmenter :: provider :: WordBreakDataV1Marker > :: load (self , req) . map (icu_provider :: DataResponse :: wrap_into_any_response) , _ => Err (icu_provider :: DataErrorKind :: MissingDataKey . with_req (key , req)) , } } } } } # [clippy :: msrv = "1.67"] pub struct BakedDataProvider ; impl_data_provider ! (BakedDataProvider) ;
+include ! ("segmenter_break_sentence_v1.rs.data") ; include ! ("segmenter_dictionary_auto_v1.rs.data") ; include ! ("segmenter_break_grapheme_cluster_v1.rs.data") ; include ! ("segmenter_dictionary_extended_v1.rs.data") ; include ! ("segmenter_break_line_v1.rs.data") ; include ! ("segmenter_lstm_auto_v1.rs.data") ; include ! ("segmenter_break_word_v1.rs.data") ; include ! ("segmenter_break_word_override_v1.rs.data") ; include ! ("segmenter_break_sentence_override_v1.rs.data") ; # [doc = r" Marks a type as a data provider. You can then use macros like"] # [doc = r" `impl_core_helloworld_v1` to add implementations."] # [doc = r""] # [doc = r" ```ignore"] # [doc = r" struct MyProvider;"] # [doc = r" const _: () = {"] # [doc = r#"     include!("path/to/generated/macros.rs");"#] # [doc = r"     make_provider!(MyProvider);"] # [doc = r"     impl_core_helloworld_v1!(MyProvider);"] # [doc = r" }"] # [doc = r" ```"] # [doc (hidden)] # [macro_export] macro_rules ! __make_provider { ($ name : ty) => { # [clippy :: msrv = "1.82"] impl $ name { # [allow (dead_code)] pub (crate) const MUST_USE_MAKE_PROVIDER_MACRO : () = () ; } icu_provider :: marker :: impl_data_provider_never_marker ! ($ name) ; } ; } # [doc (inline)] pub use __make_provider as make_provider ; # [allow (unused_macros)] macro_rules ! impl_data_provider { ($ provider : ty) => { make_provider ! ($ provider) ; impl_segmenter_break_sentence_v1 ! ($ provider) ; impl_segmenter_dictionary_auto_v1 ! ($ provider) ; impl_segmenter_break_grapheme_cluster_v1 ! ($ provider) ; impl_segmenter_dictionary_extended_v1 ! ($ provider) ; impl_segmenter_break_line_v1 ! ($ provider) ; impl_segmenter_lstm_auto_v1 ! ($ provider) ; impl_segmenter_break_word_v1 ! ($ provider) ; impl_segmenter_break_word_override_v1 ! ($ provider) ; impl_segmenter_break_sentence_override_v1 ! ($ provider) ; } ; }
